% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfit.R
\name{abfit_opts_v1_9_0}
\alias{abfit_opts_v1_9_0}
\title{Return a list of options for an ABfit analysis to maintain comparability with
analyses performed with version 1.9.0 (and earlier) of spant.}
\usage{
abfit_opts_v1_9_0(...)
}
\arguments{
\item{...}{arguments passed to \link{abfit_opts}.}
}
\value{
full list of options.
}
\description{
Return a list of options for an ABfit analysis to maintain comparability with
analyses performed with version 1.9.0 (and earlier) of spant.
}
