% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{acquire}
\alias{acquire}
\title{Simulate pulse sequence acquisition.}
\usage{
acquire(sys, rec_phase = 180, tol = 1e-04, detect = NULL)
}
\arguments{
\item{sys}{spin system object}

\item{rec_phase}{receiver phase in degrees}

\item{tol}{ignore resonance amplitudes below this threshold}

\item{detect}{detection nuclei}
}
\value{
a list of resonance amplitudes and frequencies
}
\description{
Simulate pulse sequence acquisition.
}
