% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{add_noise}
\alias{add_noise}
\title{Add noise to an mrs_data object.}
\usage{
add_noise(mrs_data, sd = 0.1, fd = TRUE)
}
\arguments{
\item{mrs_data}{data to add noise to.}

\item{sd}{standard deviation of the noise.}

\item{fd}{generate the noise samples in the frequency-domain (TRUE) or
time-domain (FALSE). This is required since the absolute value of the
standard deviation of noise samples changes when data is Fourier transformed.}
}
\value{
mrs_data object with additive normally distributed noise.
}
\description{
Add noise to an mrs_data object.
}
