% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{add_noise_spec_snr}
\alias{add_noise_spec_snr}
\title{Add noise to an mrs_data object.}
\usage{
add_noise_spec_snr(mrs_data, target_snr, sig_region = c(4, 0.5))
}
\arguments{
\item{mrs_data}{data to add noise to.}

\item{target_snr}{desired spectral SNR, note this assumes the input data is
noise-free, eg simulated data.}

\item{sig_region}{spectral limits to search for the strongest spectral data
point.}
}
\value{
mrs_data object with additive normally distributed noise.
}
\description{
Add noise to an mrs_data object.
}
