% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{align}
\alias{align}
\title{Align spectra to a reference frequency using a convolution based method.}
\usage{
align(mrs_data, ref_peak = 4.65, zf_factor = 2, lb = 2, max_shift = 20,
  ret_df = FALSE)
}
\arguments{
\item{mrs_data}{Data to be aligned.}

\item{ref_peak}{A reference frequency in ppm units.}

\item{zf_factor}{Zero filling factor to increase alignment resolution.}

\item{lb}{Line broadening to apply to the reference signal.}

\item{max_shift}{Maximum allowable shift in Hz.}

\item{ret_df}{Return frequency shifts in addition to aligned data (logical).}
}
\value{
Aligned data object.
}
\description{
Align spectra to a reference frequency using a convolution based method.
}
