% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{append_basis}
\alias{append_basis}
\title{Combine a pair of basis set objects.}
\usage{
append_basis(basis_a, basis_b)
}
\arguments{
\item{basis_a}{first basis.}

\item{basis_b}{second basis.}
}
\value{
combined basis set object.
}
\description{
Combine a pair of basis set objects.
}
