% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{append_dyns}
\alias{append_dyns}
\title{Append MRS data across the dynamic dimension.}
\usage{
append_dyns(...)
}
\arguments{
\item{...}{MRS data objects.}
}
\value{
A single MRS data object with the input objects concatenated together.
}
\description{
Append MRS data across the dynamic dimension.
}
