% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{append_regs}
\alias{append_regs}
\title{Append multiple regressor data frames into a single data frame.}
\usage{
append_regs(...)
}
\arguments{
\item{...}{input regressor data frames.}
}
\value{
output regressor data frame.
}
\description{
Append multiple regressor data frames into a single data frame.
}
