% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{apply_pulse}
\alias{apply_pulse}
\title{Simulate an RF pulse on a single spin.}
\usage{
apply_pulse(sys, rho, spin_n, angle, nuc, xy)
}
\arguments{
\item{sys}{spin system object.}

\item{rho}{density matrix.}

\item{spin_n}{spin index.}

\item{angle}{RF flip angle in degrees.}

\item{nuc}{nucleus influenced by the pulse.}

\item{xy}{x or y pulse.}
}
\value{
density matrix.
}
\description{
Simulate an RF pulse on a single spin.
}
