% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{apply_water_scaling}
\alias{apply_water_scaling}
\title{Apply water reference scaling to a fitting results object to yield metabolite
quantities in units of mM.}
\usage{
apply_water_scaling(result, ref_data, w_att = 0.7, w_conc = 35880)
}
\arguments{
\item{result}{A result object generated from fitting.}

\item{ref_data}{Water reference MRS data object.}

\item{w_att}{Water attenuation factor (default = 0.7).}

\item{w_conc}{Assumed water concentration (default = 35880).}
}
\value{
A result object with a results_ws data table added.
}
\description{
Apply water reference scaling to a fitting results object to yield metabolite
quantities in units of mM.
}
