% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Arg.mrs_data}
\alias{Arg.mrs_data}
\title{Apply Arg operator to an MRS dataset.}
\usage{
\method{Arg}{mrs_data}(z)
}
\arguments{
\item{z}{MRS data.}
}
\value{
MRS data following Arg operator.
}
\description{
Apply Arg operator to an MRS dataset.
}
