% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{array2mrs_data}
\alias{array2mrs_data}
\title{Convert a 7 dimensional array in into a mrs_data object. The array dimensions
should be ordered as : dummy, X, Y, Z, dynamic, coil, FID.}
\usage{
array2mrs_data(
  data_array,
  mrs_data = NULL,
  fs = NULL,
  ft = NULL,
  ref = NULL,
  nuc = NULL,
  fd = FALSE
)
}
\arguments{
\item{data_array}{7d data array.}

\item{mrs_data}{example data to copy acquisition parameters from.}

\item{fs}{sampling frequency in Hz.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{nuc}{nucleus that is resonant at the transmitter frequency.}

\item{fd}{flag to indicate if the matrix is in the frequency domain (logical).}
}
\value{
mrs_data object.
}
\description{
Convert a 7 dimensional array in into a mrs_data object. The array dimensions
should be ordered as : dummy, X, Y, Z, dynamic, coil, FID.
}
