% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{assign_dyns}
\alias{assign_dyns}
\title{Overwrite a subset of dynamic scans.}
\usage{
assign_dyns(mrs_data, subset, mrs_data_new)
}
\arguments{
\item{mrs_data}{dynamic MRS data to be updated.}

\item{subset}{vector containing indices to the dynamic scans to be
assigned.}

\item{mrs_data_new}{dynamic MRS data with new values.}
}
\value{
MRS data with updated dynamics.
}
\description{
Overwrite a subset of dynamic scans.
}
