% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{auto_pad_seq}
\alias{auto_pad_seq}
\title{Automatically zero pad a numerical vector.}
\usage{
auto_pad_seq(x, min_pad = 2)
}
\arguments{
\item{x}{numeric or integer vector.}

\item{min_pad}{minimum number of zeros to pad.}
}
\value{
character vector of zero-padded values.
}
\description{
Automatically zero pad a numerical vector.
}
