% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{back_extrap}
\alias{back_extrap}
\title{Back extrapolate time-domain points using the Burg autoregressive model}
\usage{
back_extrap(mrs_data, n_pts)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{n_pts}{number of points to extrapolate.}
}
\value{
back extrapolated data.
}
\description{
Back extrapolate time-domain points using the Burg autoregressive model
}
