% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{basis2dyn_mrs_data}
\alias{basis2dyn_mrs_data}
\title{Convert a basis object to a dynamic mrs_data object.}
\usage{
basis2dyn_mrs_data(basis, amps, tr)
}
\arguments{
\item{basis}{basis set object.}

\item{amps}{a data frame with each column corresponding to a basis element
and each row corresponding to each dynamic scan.}

\item{tr}{the dataset repetition time in seconds.}
}
\value{
a dynamic mrs_data object.
}
\description{
Convert a basis object to a dynamic mrs_data object.
}
