% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{basis2mrs_data}
\alias{basis2mrs_data}
\title{Convert a basis object to an mrs_data object - where basis signals are spread
across the dynamic dimension.}
\usage{
basis2mrs_data(basis)
}
\arguments{
\item{basis}{Basis set object.}
}
\value{
An mrs_data object with basis signals spread across the dynamic dimension.
}
\description{
Convert a basis object to an mrs_data object - where basis signals are spread
across the dynamic dimension.
}
