% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_als}
\alias{bc_als}
\title{Baseline correction using the ALS method.}
\usage{
bc_als(mrs_data, lambda = 10000, p = 0.001, ret_bc_only = TRUE)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{lambda}{controls the baseline flexibility.}

\item{p}{controls the penalty for negative data points.}

\item{ret_bc_only}{return the baseline corrected data only. When FALSE the
baseline estimate and input data will be returned.}
}
\value{
baseline corrected data.
}
\description{
Eilers P. H. C. and Boelens H. F. M. (2005) Baseline correction with
asymmetric least squares smoothing. Leiden Univ. Medical Centre Report.
}
