% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_constant}
\alias{bc_constant}
\title{Remove a constant baseline offset based on a reference spectral region.}
\usage{
bc_constant(mrs_data, xlim = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{xlim}{spectral range containing a flat baseline region to measure the
offset.}
}
\value{
baseline corrected data.
}
\description{
Remove a constant baseline offset based on a reference spectral region.
}
