% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_gauss}
\alias{bc_gauss}
\title{Apply and subtract a Gaussian smoother in the spectral domain.}
\usage{
bc_gauss(mrs_data, smo_ppm_sd)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{smo_ppm_sd}{Gaussian smoother sd in ppm units.}
}
\value{
smoother subtracted data.
}
\description{
Apply and subtract a Gaussian smoother in the spectral domain.
}
