% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_poly}
\alias{bc_poly}
\title{Fit and subtract a polynomial to each spectrum in a dataset.}
\usage{
bc_poly(mrs_data, p_deg = 1, fd = TRUE)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{p_deg}{polynomial degree.}

\item{fd}{perform correction in the frequency domain if TRUE, or time-domain
if FALSE. Default is to perform in the frequency domain.}
}
\value{
polynomial subtracted data.
}
\description{
Fit and subtract a polynomial to each spectrum in a dataset.
}
