% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_spline}
\alias{bc_spline}
\title{Fit and subtract a smoothing spline to each spectrum in a dataset.}
\usage{
bc_spline(mrs_data, spar = 0.5, nknots = 100)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{spar}{smoothing parameter typically between 0 and 1.}

\item{nknots}{number of spline knots.}
}
\value{
smoothing spline subtracted data.
}
\description{
Fit and subtract a smoothing spline to each spectrum in a dataset.
}
