% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{beta2lw}
\alias{beta2lw}
\title{Covert a beta value in the time-domain to an equivalent linewidth in Hz:
x * exp(-i * t * t * beta)}
\usage{
beta2lw(beta)
}
\arguments{
\item{beta}{beta damping value}
}
\value{
linewidth value in Hz
}
\description{
Covert a beta value in the time-domain to an equivalent linewidth in Hz:
x * exp(-i * t * t * beta)
}
