% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bin_spec}
\alias{bin_spec}
\title{Bin equally spaced spectral regions.}
\usage{
bin_spec(mrs_data, width = 0.05, unit = "ppm")
}
\arguments{
\item{mrs_data}{data to be "binned".}

\item{width}{bin width.}

\item{unit}{bin width unit, can be "ppm" (default) or "pts".}
}
\value{
binned mrs_data object.
}
\description{
Bin equally spaced spectral regions.
}
