% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{calc_basis_crlbs}
\alias{calc_basis_crlbs}
\title{Estimate the CRLB for each element in a basis set.}
\usage{
calc_basis_crlbs(
  basis,
  xlim = c(4, 0.2),
  zf = TRUE,
  sd = 1,
  bl_comp_pppm = NULL
)
}
\arguments{
\item{basis}{basis_set object.}

\item{xlim}{spectral range to use in ppm.}

\item{zf}{zero-fill the basis set.}

\item{sd}{standard deviation of the noise.}

\item{bl_comp_pppm}{number spline baseline components to append per-ppm.}
}
\value{
a vector of predicted errors.
}
\description{
Estimate the CRLB for each element in a basis set.
}
