% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{calc_coil_noise_cor}
\alias{calc_coil_noise_cor}
\title{Calculate the noise correlation between coil elements.}
\usage{
calc_coil_noise_cor(noise_data)
}
\arguments{
\item{noise_data}{\code{mrs_data} object with one FID for each coil element.}
}
\value{
correlation matrix.
}
\description{
Calculate the noise correlation between coil elements.
}
