% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{calc_coil_noise_sd}
\alias{calc_coil_noise_sd}
\title{Calculate the noise stanard deviation for each coil element.}
\usage{
calc_coil_noise_sd(noise_data)
}
\arguments{
\item{noise_data}{\code{mrs_data} object with one FID for each coil element.}
}
\value{
array of standard deviations.
}
\description{
Calculate the noise stanard deviation for each coil element.
}
