% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{calc_design_efficiency}
\alias{calc_design_efficiency}
\title{Calculate the efficiency of a regressor data frame.}
\usage{
calc_design_efficiency(regressor_df, contrasts)
}
\arguments{
\item{regressor_df}{input regressor data frame.}

\item{contrasts}{a vector of contrast values.}
}
\description{
Calculate the efficiency of a regressor data frame.
}
