% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfit.R
\name{calc_ed_from_lambda}
\alias{calc_ed_from_lambda}
\title{Calculate the effective dimensions of a spline smoother from lambda.}
\usage{
calc_ed_from_lambda(spline_basis, deriv_mat, lambda)
}
\arguments{
\item{spline_basis}{spline basis.}

\item{deriv_mat}{derivative matrix.}

\item{lambda}{smoothing parameter.}
}
\value{
the effective dimension value.
}
\description{
Calculate the effective dimensions of a spline smoother from lambda.
}
