% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{calc_peak_info}
\alias{calc_peak_info}
\title{Search for the highest peak in a spectral region and return the frequency,
height and FWHM.}
\usage{
calc_peak_info(mrs_data, xlim = c(4, 0.5), interp_f = 4)
}
\arguments{
\item{mrs_data}{an object of class \code{mrs_data}.}

\item{xlim}{frequency range in ppm to search for the highest peak.}

\item{interp_f}{interpolation factor, defults to 4x.}
}
\value{
list of highest peak frequency, hight and FWHM.
}
\description{
Search for the highest peak in a spectral region and return the frequency,
height and FWHM.
}
