% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{calc_peak_info_vec}
\alias{calc_peak_info_vec}
\title{Calculate the FWHM of a peak from a vector of intensity values.}
\usage{
calc_peak_info_vec(data_pts, interp_f)
}
\arguments{
\item{data_pts}{input vector.}

\item{interp_f}{interpolation factor to improve the FWHM estimate.}
}
\value{
a vector of: x position of the highest data point, maximum peak
value in the y axis, FWHM in the units of data points.
}
\description{
Calculate the FWHM of a peak from a vector of intensity values.
}
