% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_sd_poly}
\alias{calc_sd_poly}
\title{Perform a polynomial fit, subtract and return the standard deviation of the
residuals.}
\usage{
calc_sd_poly(y, degree = 1)
}
\arguments{
\item{y}{an array.}

\item{degree}{polynomial degree.}
}
\value{
the standard deviation of the fit residuals.
}
\description{
Perform a polynomial fit, subtract and return the standard deviation of the
residuals.
}
