% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{check_spant_deps_dir}
\alias{check_spant_deps_dir}
\title{Check the "spant_deps" directory exists in the users home folder.}
\usage{
check_spant_deps_dir()
}
\description{
Check the "spant_deps" directory exists in the users home folder.
}
