% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circ_mask}
\alias{circ_mask}
\title{Create a logical circular mask spanning the full extent of the d x d matrix.}
\usage{
circ_mask(d)
}
\arguments{
\item{d}{diameter of the mask.}
}
\value{
logical d x d mask matrix.
}
\description{
Create a logical circular mask spanning the full extent of the d x d matrix.
}
