% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{coherence_filter}
\alias{coherence_filter}
\title{Zero all coherence orders other than the one supplied as an argument.}
\usage{
coherence_filter(sys, rho, order = 0)
}
\arguments{
\item{sys}{spin system object.}

\item{rho}{density matrix.}

\item{order}{coherence order to keep (default is 0).}
}
\value{
density matrix.
}
\description{
Zero all coherence orders other than the one supplied as an argument.
}
