% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{comb_coils_fp_pc}
\alias{comb_coils_fp_pc}
\title{Combine coil data following phase correction based on the first data point
in the FID.}
\usage{
comb_coils_fp_pc(metab, ref = NULL, sum_coils = TRUE,
  ret_ref = FALSE)
}
\arguments{
\item{metab}{MRS data containing metabolite data.}

\item{ref}{MRS data containing reference data (optional).}

\item{sum_coils}{sum the coil elements as a final step (logical).}

\item{ret_ref}{return the reference data following correction.}
}
\value{
MRS data.
}
\description{
Combine coil data following phase correction based on the first data point
in the FID.
}
