% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{comb_coils_mrsi_gls}
\alias{comb_coils_mrsi_gls}
\title{Combine MRSI coil data using the GLS method presented by An et al
JMRI 37:1445-1450 (2013).}
\usage{
comb_coils_mrsi_gls(metab, noise_pts = 30, noise_mrs = NULL)
}
\arguments{
\item{metab}{MRSI data containing metabolite data.}

\item{noise_pts}{number of points from the end of the FIDs to use for noise
covariance estimation.}

\item{noise_mrs}{MRS data containing noise information for each coil.}
}
\value{
coil combined MRSI data.
}
\description{
Combine MRSI coil data using the GLS method presented by An et al
JMRI 37:1445-1450 (2013).
}
