% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{comb_coils_svs_gls}
\alias{comb_coils_svs_gls}
\title{Combine SVS coil data using the GLS method presented by An et al
JMRI 37:1445-1450 (2013).}
\usage{
comb_coils_svs_gls(
  metab,
  ref = NULL,
  noise_pts = 256,
  noise_mrs = NULL,
  use_mean_sens = TRUE,
  use_ref_sens = FALSE,
  change_order = NULL,
  no_noise_corr = FALSE,
  bc_poly_noise = 2
)
}
\arguments{
\item{metab}{MRS data containing metabolite data.}

\item{ref}{MRS data containing reference data (optional).}

\item{noise_pts}{number of points from the end of the FIDs to use for noise
covariance estimation.}

\item{noise_mrs}{MRS data containing noise information for each coil.}

\item{use_mean_sens}{use the dynamic mean to estimate coil sensitivities.}

\item{use_ref_sens}{use the reference data to estimate coil sensitivities.}

\item{change_order}{change the order of the noise matrix reshaping, for
testing purposes only.}

\item{no_noise_corr}{assume noise is uncorrelated between coils, for testing
purposes only. Defaults to FALSE.}

\item{bc_poly_noise}{baseline correct the noise samples with a polynomial in
the time-domain. Defaults to 2, which performs a second-order polynomial
correction. Set to NULL to disable.}
}
\value{
coil combined MRS data.
}
\description{
Combine SVS coil data using the GLS method presented by An et al
JMRI 37:1445-1450 (2013).
}
