% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{comb_coils_wref}
\alias{comb_coils_wref}
\title{Combine coil data based on the first data point of a reference signal.}
\usage{
comb_coils_wref(metab_mrs, ref_mrs, scale = TRUE)
}
\arguments{
\item{metab_mrs}{MRS data containing metabolite data.}

\item{ref_mrs}{MRS data containing reference data.}

\item{scale}{Option to rescale coil elements based on the first data point
(logical).}
}
\value{
Data with coil elements combined.
}
\description{
Elements are phased and optionally scaled prior to summation.
}
