% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{comb_fit_list_dyn}
\alias{comb_fit_list_dyn}
\title{Combine a list of fit_result objects into a single fit object containing a
dynamic series.}
\usage{
comb_fit_list_dyn(fit_list)
}
\arguments{
\item{fit_list}{list of fit_result objects.}
}
\value{
fit_result object.
}
\description{
Combine a list of fit_result objects into a single fit object containing a
dynamic series.
}
