% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{comb_fit_list_dyns}
\alias{comb_fit_list_dyns}
\title{Combine a list of fit results into a single fit result object across the
dynamic dimension.}
\usage{
comb_fit_list_dyns(fit_list, dyn_basis_sets = FALSE)
}
\arguments{
\item{fit_list}{a list of fit_result objects.}

\item{dyn_basis_sets}{save all basis sets separately  as a list, defaults to
FALSE.}
}
\value{
fit result object.
}
\description{
Combine a list of fit results into a single fit result object across the
dynamic dimension.
}
