% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{comb_fit_list_fit_tables}
\alias{comb_fit_list_fit_tables}
\title{Combine all fitting data points from a list of fits into a single dataframe.}
\usage{
comb_fit_list_fit_tables(fit_list)
}
\arguments{
\item{fit_list}{list of fit_result objects.}
}
\value{
a dataframe containing the fit data points.
}
\description{
Combine all fitting data points from a list of fits into a single dataframe.
}
