% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{comb_fit_list_result_tables}
\alias{comb_fit_list_result_tables}
\title{Combine the fit result tables from a list of fit results.}
\usage{
comb_fit_list_result_tables(fit_list, add_extra = TRUE)
}
\arguments{
\item{fit_list}{a list of fit_result objects.}

\item{add_extra}{add variables in the extra data frame to the output (TRUE).}
}
\value{
a data frame combine all fit result tables with an additional id
column to differentiate between data sets. Any variables in the extra data
frame may be optionally added to the result.
}
\description{
Combine the fit result tables from a list of fit results.
}
