% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{comb_fit_tables}
\alias{comb_fit_tables}
\title{Combine all fitting data points into a single data frame.}
\usage{
comb_fit_tables(fit_res)
}
\arguments{
\item{fit_res}{a single fit_result object.}
}
\value{
a data frame containing the fit data points.
}
\description{
Combine all fitting data points into a single data frame.
}
