% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{combine_metab_ref}
\alias{combine_metab_ref}
\title{Combine a reference and metabolite mrs_data object.}
\usage{
combine_metab_ref(metab, ref)
}
\arguments{
\item{metab}{metabolite mrs_data object.}

\item{ref}{reference mrs_data object.}
}
\value{
Combined metabolite and reference mrs_data object.
}
\description{
Combine a reference and metabolite mrs_data object.
}
