% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{combine_result_tables}
\alias{combine_result_tables}
\title{Combine the fit result tables from a list of fit results.}
\usage{
combine_result_tables(fit_res_list)
}
\arguments{
\item{fit_res_list}{a list of fit_result objects.}
}
\value{
a data frame combine all fit result tables with an additional id
column to differentiate between data sets. Any variables in the extra data
frame will also be added to the result.
}
\description{
Combine the fit result tables from a list of fit results.
}
