% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{conv_mrs}
\alias{conv_mrs}
\title{Convolve two MRS data objects.}
\usage{
conv_mrs(mrs_data, conv)
}
\arguments{
\item{mrs_data}{MRS data to be convolved.}

\item{conv}{convolution data stored as an mrs_data object.}
}
\value{
convolved data.
}
\description{
Convolve two MRS data objects.
}
