% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{crop_basis}
\alias{crop_basis}
\title{Crop \code{basis_set} object based on a frequency range.}
\usage{
crop_basis(basis, xlim = c(4, 0.2), scale = "ppm")
}
\arguments{
\item{basis}{basis_set object to be cropped in the spectral dimension.}

\item{xlim}{range of values to crop in the spectral dimension eg
xlim = c(4, 0.2).}

\item{scale}{the units to use for the frequency scale, can be one of: "ppm",
"hz" or "points".}
}
\value{
cropped \code{mrs_data} object.
}
\description{
Crop \code{basis_set} object based on a frequency range.
}
