% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{crop_td_pts}
\alias{crop_td_pts}
\title{Crop \code{mrs_data} object data points in the time-domain.}
\usage{
crop_td_pts(mrs_data, start = NULL, end = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{start}{starting data point (defaults to 1).}

\item{end}{ending data point (defaults to the last saved point).}
}
\value{
cropped \code{mrs_data} object.
}
\description{
Crop \code{mrs_data} object data points in the time-domain.
}
