% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{crop_td_pts_end}
\alias{crop_td_pts_end}
\title{Crop \code{mrs_data} object data points at the end of the FID.}
\usage{
crop_td_pts_end(mrs_data, pts)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{pts}{number of points to remove from the end of the FID.}
}
\value{
cropped \code{mrs_data} object.
}
\description{
Crop \code{mrs_data} object data points at the end of the FID.
}
