% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{crop_xy}
\alias{crop_xy}
\title{Crop an MRSI dataset in the x-y direction}
\usage{
crop_xy(mrs_data, x_dim, y_dim)
}
\arguments{
\item{mrs_data}{MRS data object.}

\item{x_dim}{x dimension output length.}

\item{y_dim}{y dimension output length.}
}
\value{
selected subset of MRS data.
}
\description{
Crop an MRSI dataset in the x-y direction
}
