% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crossprod_3d}
\alias{crossprod_3d}
\title{Compute the vector cross product between vectors x and y. Adapted from
http://stackoverflow.com/questions/15162741/what-is-rs-crossproduct-function}
\usage{
crossprod_3d(x, y)
}
\arguments{
\item{x}{vector of length 3.}

\item{y}{vector of length 3.}
}
\value{
vector cross product of x and y.
}
\description{
Compute the vector cross product between vectors x and y. Adapted from
http://stackoverflow.com/questions/15162741/what-is-rs-crossproduct-function
}
