% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{decimate_mrs_fd}
\alias{decimate_mrs_fd}
\title{Decimate an MRS signal to half the original sampling frequency by filtering
in the frequency domain before down sampling.}
\usage{
decimate_mrs_fd(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
decimated data at half the original sampling frequency.
}
\description{
Decimate an MRS signal to half the original sampling frequency by filtering
in the frequency domain before down sampling.
}
