% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{deconv_mrs}
\alias{deconv_mrs}
\title{Deconvolve two MRS data objects.}
\usage{
deconv_mrs(mrs_data_a, mrs_data_b)
}
\arguments{
\item{mrs_data_a}{MRS data to be deconvolved.}

\item{mrs_data_b}{MRS data to be deconvolved.}
}
\value{
deconvolved data.
}
\description{
Deconvolve two MRS data objects.
}
