% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{def_acq_paras}
\alias{def_acq_paras}
\title{Return (and optionally modify using the input arguments) a list of the
default acquisition parameters.}
\usage{
def_acq_paras(ft = getOption("spant.def_ft"),
  fs = getOption("spant.def_fs"), N = getOption("spant.def_N"),
  ref = getOption("spant.def_ref"))
}
\arguments{
\item{ft}{specify the transmitter frequency in Hz.}

\item{fs}{specify the sampling frequency in Hz.}

\item{N}{specify the number of data points in the spectral dimension.}

\item{ref}{specify the reference value for ppm scale.}
}
\value{
A list containing the following elements:
\itemize{
\item ft Transmitter frequency in Hz.
\item fs Sampling frequency in Hz.
\item N Number of data points in the spectral dimension.
\item ref Reference value for ppm scale.
}
}
\description{
Return (and optionally modify using the input arguments) a list of the
default acquisition parameters.
}
