% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{def_N}
\alias{def_N}
\title{Return the default number of data points in the spectral dimension.}
\usage{
def_N()
}
\value{
number of data points in the spectral dimension.
}
\description{
Return the default number of data points in the spectral dimension.
}
