% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{def_ref}
\alias{def_ref}
\title{Return the default reference value for ppm scale.}
\usage{
def_ref()
}
\value{
reference value for ppm scale.
}
\description{
Return the default reference value for ppm scale.
}
