% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{diff.mrs_data}
\alias{diff.mrs_data}
\title{Apply the diff operator to an MRS dataset.}
\usage{
\method{diff}{mrs_data}(x, ...)
}
\arguments{
\item{x}{MRS data.}

\item{...}{additional arguments to the diff function.}
}
\value{
MRS data following diff operator.
}
\description{
Apply the diff operator to an MRS dataset.
}
