% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{diff_mrs}
\alias{diff_mrs}
\title{Apply the diff operator to an MRS dataset in the FID/spectral dimension.}
\usage{
diff_mrs(mrs_data, ...)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{...}{additional arguments to the diff function.}
}
\value{
MRS data following diff operator.
}
\description{
Apply the diff operator to an MRS dataset in the FID/spectral dimension.
}
