% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{downsample_mrs_fd}
\alias{downsample_mrs_fd}
\title{Downsample an MRS signal by a factor of 2 using an FFT "brick-wall" filter.}
\usage{
downsample_mrs_fd(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
downsampled data.
}
\description{
Downsample an MRS signal by a factor of 2 using an FFT "brick-wall" filter.
}
