% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{downsample_mrs_td}
\alias{downsample_mrs_td}
\title{Downsample an MRS signal by a factor of 2 by removing every other data point
in the time-domain. Note, signals outside the new sampling frequency will be
aliased.}
\usage{
downsample_mrs_td(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
downsampled data.
}
\description{
Downsample an MRS signal by a factor of 2 by removing every other data point
in the time-domain. Note, signals outside the new sampling frequency will be
aliased.
}
