% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{dyn_acq_times}
\alias{dyn_acq_times}
\title{Return a time scale vector of acquisition times for a dynamic MRS scan. The
first temporal scan is assigned a value of 0.}
\usage{
dyn_acq_times(mrs_data = NULL, tr = NULL, Ndyns = NULL, Ntrans = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{tr}{repetition time.}

\item{Ndyns}{number of dynamic scans stored, potentially less than Ntrans
if block averaging has been performed.}

\item{Ntrans}{number of dynamic scans acquired.}
}
\value{
time scale vector in units of seconds.
}
\description{
Return a time scale vector of acquisition times for a dynamic MRS scan. The
first temporal scan is assigned a value of 0.
}
