% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{dyns}
\alias{dyns}
\title{Return the number of dynamic scans in an MRS dataset.}
\usage{
dyns(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
Number of dynamic scans
}
\description{
Return the number of dynamic scans in an MRS dataset.
}
