% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{ecc}
\alias{ecc}
\title{Apply eddy current correction using the Klose method:}
\usage{
ecc(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data including a reference dataset.}
}
\value{
Corrected data.
}
\description{
In vivo proton spectroscopy in presence of eddy currents.
Klose U.
Magn Reson Med. 1990 Apr;14(1):26-30.
}
