% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elliptical_mask}
\alias{elliptical_mask}
\title{Create an elliptical mask stored as a matrix of logical values.}
\usage{
elliptical_mask(xN, yN, x0, y0, xr, yr, angle)
}
\arguments{
\item{xN}{number of pixels in the x dimension.}

\item{yN}{number of pixels in the y dimension.}

\item{x0}{centre of ellipse in the x direction in units of pixels.}

\item{y0}{centre of ellipse in the y direction in units of pixels.}

\item{xr}{radius in the x direction in units of pixels.}

\item{yr}{radius in the y direction in units of pixels.}

\item{angle}{angle of rotation in degrees.}
}
\value{
logical mask matrix with dimensions fov_yN x fov_xN.
}
\description{
Create an elliptical mask stored as a matrix of logical values.
}
