% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{est_noise_sd}
\alias{est_noise_sd}
\title{Estimate the standard deviation of the noise from a segment of an mrs_data object.}
\usage{
est_noise_sd(mrs_data, n = 100, offset = 100, p_order = 2)
}
\arguments{
\item{mrs_data}{MRS data object.}

\item{n}{number of data points (taken from the end of array) to use in the estimation.}

\item{offset}{number of final points to exclude from the calculation.}

\item{p_order}{polynomial order to fit to the data before estimating the standard deviation.}
}
\value{
standard deviation array.
}
\description{
Estimate the standard deviation of the noise from a segment of an mrs_data object.
}
