% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{faceoff}
\alias{faceoff}
\title{Deface a T1 weighted head scan using the FaceOff method described in :
https://github.com/srikash/FaceOff. Requires ANTs to be installed.}
\usage{
faceoff(mri_path, out_dir = NULL)
}
\arguments{
\item{mri_path}{path to the volumetric T1 data.}

\item{out_dir}{optional output directory. Defaults to the same directory
as mri_path if not specified.}
}
\description{
Deface a T1 weighted head scan using the FaceOff method described in :
https://github.com/srikash/FaceOff. Requires ANTs to be installed.
}
