% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fd2td}
\alias{fd2td}
\title{Transform frequency-domain data to the time-domain.}
\usage{
fd2td(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data in frequency-domain representation.}
}
\value{
MRS data in time-domain representation.
}
\description{
Transform frequency-domain data to the time-domain.
}
