% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fd_gauss_smo}
\alias{fd_gauss_smo}
\title{Apply a Gaussian smoother in the spectral domain.}
\usage{
fd_gauss_smo(mrs_data, smo_ppm_sd)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{smo_ppm_sd}{Gaussian smoother sd in ppm units.}
}
\value{
spectrally smoothed data.
}
\description{
Apply a Gaussian smoother in the spectral domain.
}
