% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fft_dyns}
\alias{fft_dyns}
\title{Perform a Fourier Transform over the dynamic MRS data dimension.}
\usage{
fft_dyns(mrs_data, ret_mod = FALSE)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{ret_mod}{return the modulus of the FT.}
}
\value{
Fourier Transformed data.
}
\description{
Perform a Fourier Transform over the dynamic MRS data dimension.
}
