% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{find_bids_mrs}
\alias{find_bids_mrs}
\title{Search for MRS data files in a BIDS filesystem structure.}
\usage{
find_bids_mrs(path)
}
\arguments{
\item{path}{path to the directory containing the BIDS structure.}
}
\value{
data frame containing full paths and information on each MRS file.
}
\description{
Search for MRS data files in a BIDS filesystem structure.
}
