% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_mrs_files}
\alias{find_mrs_files}
\title{Find valid MRS data files recursively from a directory path.}
\usage{
find_mrs_files(dir)
}
\arguments{
\item{dir}{a directory path.}
}
\value{
a vector of valid MRS data files.
}
\description{
Find valid MRS data files recursively from a directory path.
}
