% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fit_amps}
\alias{fit_amps}
\title{Extract the fit amplitudes from an object of class \code{fit_result}.}
\usage{
fit_amps(x, inc_index = FALSE, sort_names = TRUE,
  append_common_1h_comb = TRUE)
}
\arguments{
\item{x}{\code{fit_result} object.}

\item{inc_index}{include columns for the voxel index.}

\item{sort_names}{sort the basis set names alphabetically.}

\item{append_common_1h_comb}{append commonly used 1H metabolite combinations
eg TNAA = NAA + NAAG.}
}
\value{
a dataframe of amplitudes.
}
\description{
Extract the fit amplitudes from an object of class \code{fit_result}.
}
