% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{fit_diags}
\alias{fit_diags}
\title{Calculate diagnostic information for object of class \code{fit_result}}
\usage{
fit_diags(x, amps = NULL)
}
\arguments{
\item{x}{\code{fit_result} object}

\item{amps}{known metabolite amplitudes}
}
\value{
a dataframe of diagnostic information
}
\description{
Calculate diagnostic information for object of class \code{fit_result}
}
