% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{fit_res2csv}
\alias{fit_res2csv}
\title{Write fit results table to a csv file.}
\usage{
fit_res2csv(fit_res, fname, unscaled = FALSE)
}
\arguments{
\item{fit_res}{fit result object.}

\item{fname}{filename of csv file.}

\item{unscaled}{output the unscaled result table (default = FALSE).}
}
\description{
Write fit results table to a csv file.
}
