% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svs.R
\name{fit_svs_btrg_v1}
\alias{fit_svs_btrg_v1}
\title{A simple wrapper of fit_svs with defaults adjusted to improve the analysis
of lower quality brain tumour spectra.}
\usage{
fit_svs_btrg_v1(
  append_basis = c("peth", "cit", "gly"),
  pre_align_ref_freq = c(4.65),
  hsvd_width = 50,
  fit_opts = NULL,
  ...
)
}
\arguments{
\item{append_basis}{names of extra signals to add to the default basis. Eg
append_basis = c("peth", "cit"). Use get_mol_names() function to print all
available signals. Cannot be used with precompiled basis sets.}

\item{pre_align_ref_freq}{reference frequency in ppm units. More than one
frequency may be specified. Defaults to : c(4.65).}

\item{hsvd_width}{set the width of the HSVD filter in Hz. Note the applied
width is between -width and +width Hz, with 0 Hz being defined at the centre
of the spectral width. Default is set to 50 Hz.}

\item{fit_opts}{options to pass to the fitting method.}

\item{...}{other arguments to pass to fit_svs.}
}
\description{
A simple wrapper of fit_svs with defaults adjusted to improve the analysis
of lower quality brain tumour spectra.
}
