% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{fit_tab2csv}
\alias{fit_tab2csv}
\title{Write fit results table to a csv file.}
\usage{
fit_tab2csv(x, fname, pvc = FALSE)
}
\arguments{
\item{x}{fit results table.}

\item{fname}{filename of csv file.}

\item{pvc}{output PVC or raw results (logical).}
}
\description{
Write fit results table to a csv file.
}
