% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{fmrs_dataset_spec_glm}
\alias{fmrs_dataset_spec_glm}
\title{Perform spectral GLM analysis of an fMRS dataset.}
\usage{
fmrs_dataset_spec_glm(
  regressor_df,
  analysis_dir = "spant_analysis",
  exclude_labels = NULL,
  labels = NULL
)
}
\arguments{
\item{regressor_df}{a data frame containing temporal regressors to be applied
to each spectral datapoint.}

\item{analysis_dir}{directory containing preprocessed data generated by
the preproc_svs_dataset function.}

\item{exclude_labels}{vector of labels of scans to exclude, eg poor quality}

\item{labels}{labels to describe each data set.}
}
\description{
Perform spectral GLM analysis of an fMRS dataset.
}
