% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fp_phase}
\alias{fp_phase}
\title{Return the phase of the first data point in the time-domain.}
\usage{
fp_phase(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
phase values in degrees.
}
\description{
Return the phase of the first data point in the time-domain.
}
