% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fp_phase_correct}
\alias{fp_phase_correct}
\title{Perform a zeroth order phase correction based on the phase of the first data
point in the time-domain.}
\usage{
fp_phase_correct(mrs_data, ret_phase = FALSE)
}
\arguments{
\item{mrs_data}{MRS data to be corrected.}

\item{ret_phase}{Return phase values (logical).}
}
\value{
Corrected data or a list with corrected data and optional phase
values.
}
\description{
Perform a zeroth order phase correction based on the phase of the first data
point in the time-domain.
}
