% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fp_scale}
\alias{fp_scale}
\title{Scale the first time-domain data point in an mrs_data object.}
\usage{
fp_scale(mrs_data, scale = 0.5)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{scale}{scaling value, defaults to 0.5.}
}
\value{
scaled mrs_data object.
}
\description{
Scale the first time-domain data point in an mrs_data object.
}
