% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{fp_scale_basis}
\alias{fp_scale_basis}
\title{Scale the first time-domain data point in an basis object.}
\usage{
fp_scale_basis(basis, scale = 0.5)
}
\arguments{
\item{basis}{basis object.}

\item{scale}{scaling value, defaults to 0.5.}
}
\value{
scaled basis object.
}
\description{
Scale the first time-domain data point in an basis object.
}
