% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{fs}
\alias{fs}
\title{Return the sampling frequency in Hz of an MRS dataset.}
\usage{
fs(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
Sampling frequency in Hz.
}
\description{
Return the sampling frequency in Hz of an MRS dataset.
}
