% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{ft_dyns}
\alias{ft_dyns}
\title{Apply the Fourier transform over the dynamic dimension.}
\usage{
ft_dyns(mrs_data, ft_shift = FALSE, ret_mod = FALSE)
}
\arguments{
\item{mrs_data}{MRS data where the dynamic dimension is in the time-domain.}

\item{ft_shift}{apply FT shift to the output, default is FALSE.}

\item{ret_mod}{return the modulus out the transform, default is FALSE.}
}
\value{
transformed MRS data.
}
\description{
Apply the Fourier transform over the dynamic dimension.
}
