% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ft_shift}
\alias{ft_shift}
\title{Perform a fft and ffshift on a vector.}
\usage{
ft_shift(vec_in)
}
\arguments{
\item{vec_in}{vector input.}
}
\value{
output vector.
}
\description{
Perform a fft and ffshift on a vector.
}
