% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ft_shift_mat}
\alias{ft_shift_mat}
\title{Perform a fft and fftshift on a matrix with each column replaced by its
shifted fft.}
\usage{
ft_shift_mat(mat_in)
}
\arguments{
\item{mat_in}{matrix input.}
}
\value{
output matrix.
}
\description{
Perform a fft and fftshift on a matrix with each column replaced by its
shifted fft.
}
