% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gausswin_2d}
\alias{gausswin_2d}
\title{Create a two dimensional Gaussian window function stored as a matrix.}
\usage{
gausswin_2d(xN, yN, x0, y0, xw, yw)
}
\arguments{
\item{xN}{number of pixels in the x dimension.}

\item{yN}{number of pixels in the y dimension.}

\item{x0}{centre of window function in the x direction in units of pixels.
Note, only integer values are applied.}

\item{y0}{centre of window function in the y direction in units of pixels.
Note, only integer values are applied.}

\item{xw}{the reciprocal of the standard deviation of the Gaussian window in
x direction.}

\item{yw}{the reciprocal of the standard deviation of the Gaussian window in
y direction.}
}
\value{
matrix with dimensions fov_yN x fov_xN.
}
\description{
Create a two dimensional Gaussian window function stored as a matrix.
}
