% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_baseline_reg}
\alias{gen_baseline_reg}
\title{Generate baseline regressor.}
\usage{
gen_baseline_reg(mrs_data)
}
\arguments{
\item{mrs_data}{mrs_data object for timing information.}
}
\value{
a single baseline regressor with value of 1.
}
\description{
Generate baseline regressor.
}
