% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_bold_reg}
\alias{gen_bold_reg}
\title{Generate BOLD regressors.}
\usage{
gen_bold_reg(
  onset,
  duration = NULL,
  trial_type = NULL,
  mrs_data = NULL,
  tr = NULL,
  Ndyns = NULL,
  Ntrans = NULL,
  match_tr = TRUE,
  dt = 0.1,
  normalise = FALSE
)
}
\arguments{
\item{onset}{stimulus onset in seconds.}

\item{duration}{stimulus duration in seconds.}

\item{trial_type}{string label for the stimulus.}

\item{mrs_data}{mrs_data object for timing information.}

\item{tr}{repetition time.}

\item{Ndyns}{number of dynamic scans stored, potentially less than Ntrans
if block averaging has been performed.}

\item{Ntrans}{number of dynamic scans acquired.}

\item{match_tr}{match the output to the input mrs_data.}

\item{dt}{timing resolution for internal calculations.}

\item{normalise}{normalise the response function to have a maximum value of
one.}
}
\value{
BOLD regressor data frame.
}
\description{
Generate BOLD regressors.
}
