% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_bold_rf}
\alias{gen_bold_rf}
\title{Generate a BOLD response function.}
\usage{
gen_bold_rf(onset, duration, trial_type, mrs_data, match_tr = TRUE, dt = 0.1)
}
\arguments{
\item{onset}{stimulus onset in seconds.}

\item{duration}{stimulus duration in seconds.}

\item{trial_type}{string label for the stimulus.}

\item{mrs_data}{mrs_data object for timing information.}

\item{match_tr}{match the output to the input mrs_data.}

\item{dt}{timing resolution for internal calculations.}
}
\value{
BOLD response function.
}
\description{
Generate a BOLD response function.
}
