% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{gen_F}
\alias{gen_F}
\title{Generate the F product operator}
\usage{
gen_F(sys, op, detect = NULL)
}
\arguments{
\item{sys}{spin system object}

\item{op}{operator, one of "x", "y", "z", "p", "m"}

\item{detect}{detection nuclei}
}
\value{
F product operator matrix
}
\description{
Generate the F product operator
}
