% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{gen_F_xy}
\alias{gen_F_xy}
\title{Generate the Fxy product operator with a specified phase}
\usage{
gen_F_xy(sys, phase, detect = NULL)
}
\arguments{
\item{sys}{spin system object}

\item{phase}{phase angle in degrees}

\item{detect}{detection nuclei}
}
\value{
product operator matrix
}
\description{
Generate the Fxy product operator with a specified phase
}
