% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_group_reg}
\alias{gen_group_reg}
\title{Expand a regressor matrix for a group analysis.}
\usage{
gen_group_reg(regressor_df, n)
}
\arguments{
\item{regressor_df}{input regressor data frame.}

\item{n}{number of datasets n the group.}
}
\description{
Expand a regressor matrix for a group analysis.
}
