% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_impulse_reg}
\alias{gen_impulse_reg}
\title{Generate impulse regressors.}
\usage{
gen_impulse_reg(onset, trial_type = NULL, mrs_data = NULL)
}
\arguments{
\item{onset}{stimulus onset in seconds.}

\item{trial_type}{string label for the stimulus.}

\item{mrs_data}{mrs_data object for timing information.}
}
\value{
impulse regressors data frame.
}
\description{
Generate impulse regressors.
}
