% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_poly_reg}
\alias{gen_poly_reg}
\title{Generate polynomial regressors.}
\usage{
gen_poly_reg(mrs_data, degree)
}
\arguments{
\item{mrs_data}{mrs_data object for timing information.}

\item{degree}{the degree of the polynomial.}
}
\value{
polynomial regressors.
}
\description{
Generate polynomial regressors.
}
