% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_spline_reg}
\alias{gen_spline_reg}
\title{Generate spline regressors.}
\usage{
gen_spline_reg(number, mrs_data = NULL, tr = NULL, Ndyns = NULL, Ntrans = NULL)
}
\arguments{
\item{number}{the number of spline functions to generate.}

\item{mrs_data}{mrs_data object for timing information.}

\item{tr}{repetition time.}

\item{Ndyns}{number of dynamic scans stored, potentially less than Ntrans
if block averaging has been performed.}

\item{Ntrans}{number of dynamic scans acquired.}
}
\value{
spline regressors.
}
\description{
Generate spline regressors.
}
