% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{get_1h_brain_basis_names}
\alias{get_1h_brain_basis_names}
\title{Return a character vector of common 1H molecules found in healthy human
brain.}
\usage{
get_1h_brain_basis_names(inc_lip_mm = TRUE)
}
\arguments{
\item{inc_lip_mm}{include Lipid and MM basis signals.}
}
\value{
a character vector of molecule names.
}
\description{
Note, this is a basic set and PEth and Gly should be also be included for
higher quality MRS data.
}
