% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{get_1h_spectre_basis_names}
\alias{get_1h_spectre_basis_names}
\title{Return a character vector of molecules included in the Gold Star Phantoms
SPECTRE phantom.}
\usage{
get_1h_spectre_basis_names()
}
\value{
a character vector of molecule names.
}
\description{
Return a character vector of molecules included in the Gold Star Phantoms
SPECTRE phantom.
}
