% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_acq_paras}
\alias{get_acq_paras}
\title{Return acquisition parameters from a MRS data object.}
\usage{
get_acq_paras(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
A list of acquisition parameters.
}
\description{
Return acquisition parameters from a MRS data object.
}
