% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{get_ants_dir}
\alias{get_ants_dir}
\title{Return the ANTs installation directory, or throw an error if not found.}
\usage{
get_ants_dir()
}
\value{
ANTs installation directory.
}
\description{
Will check and return the "spant.ants_dir" option set by set_ants_dir. If
not set, will search the spant_resources directory for ANTs and return the most
recent version if multiple are found.
}
