% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{get_basis_subset}
\alias{get_basis_subset}
\title{Return a subset of the input basis.}
\usage{
get_basis_subset(basis, names, invert = FALSE)
}
\arguments{
\item{basis}{input basis.}

\item{names}{basis set elements to keep in the returned object.}

\item{invert}{set to true to return all basis elements except those given in
the names argument.}
}
\value{
a subset of the input basis.
}
\description{
Return a subset of the input basis.
}
