% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{get_def_acq_paras}
\alias{get_def_acq_paras}
\title{Return a list of the default acquisition parameters.}
\usage{
get_def_acq_paras()
}
\value{
A list containing the following elements:
\itemize{
\item ft Trasmitter frequency in Hz.
\item fs Sampling frequency in Hz.
\item N Number of data points in the spectral dimension.
\item ref Reference value for ppm scale.
}
}
\description{
Return a list of the default acquisition parameters.
}
