% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_dyn_av_scheme}
\alias{get_dyn_av_scheme}
\title{Average sets of dynamics according to a scheme vector.}
\usage{
get_dyn_av_scheme(mrs_data, av_scheme)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{av_scheme}{vector containing consecutive integer values (starting at 1)
representing sets of dynamics to average. This vector must have the same
length as the number of dynamic scans in mrs_data. Any elements set to NA
will not contribute to averaging.}
}
\value{
Dynamic MRS data set containing the averaged dynamic sets.
}
\description{
Average sets of dynamics according to a scheme vector.
}
