% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_dyns}
\alias{get_dyns}
\title{Extract a subset of dynamic scans.}
\usage{
get_dyns(mrs_data, subset)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{subset}{vector containing indices to the dynamic scans to be
returned.}
}
\value{
MRS data containing the subset of requested dynamics.
}
\description{
Extract a subset of dynamic scans.
}
