% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_even_dyns}
\alias{get_even_dyns}
\title{Return even numbered dynamic scans starting from 1 (2,4,6...).}
\usage{
get_even_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}
}
\value{
Dynamic MRS data containing even numbered scans.
}
\description{
Return even numbered dynamic scans starting from 1 (2,4,6...).
}
