% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{get_fit_table}
\alias{get_fit_table}
\title{Combine all fitting data points into a single dataframe.}
\usage{
get_fit_table(fit_res)
}
\arguments{
\item{fit_res}{a single fit_result object.}
}
\value{
a dataframe containing the fit data points.
}
\description{
Combine all fitting data points into a single dataframe.
}
