% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_fp}
\alias{get_fp}
\title{Return the first time-domain data point.}
\usage{
get_fp(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
first time-domain data point.
}
\description{
Return the first time-domain data point.
}
