% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_shapes.R
\name{get_gaussian_pulse}
\alias{get_gaussian_pulse}
\title{Generate a gaussian pulse shape.}
\usage{
get_gaussian_pulse(angle, n, trunc = 1)
}
\arguments{
\item{angle}{pulse angle in degrees.}

\item{n}{number of points to generate.}

\item{trunc}{percentage truncation factor.}
}
\description{
Generate a gaussian pulse shape.
}
