% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{get_hrf}
\alias{get_hrf}
\title{Generate a double gamma model of the HRF as used in SPM.}
\usage{
get_hrf(end_t = 30, res_t = 0.01)
}
\arguments{
\item{end_t}{last time point to generate in seconds.}

\item{res_t}{temporal resolution in seconds, defaults to 10ms.}
}
\value{
a data.frame of time and HRF vectors.
}
\description{
Generate a double gamma model of the HRF as used in SPM.
}
