% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_metab}
\alias{get_metab}
\title{Extract the metabolite component from an mrs_data object.}
\usage{
get_metab(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
Metabolite component.
}
\description{
Extract the metabolite component from an mrs_data object.
}
