% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_parameters.R
\name{get_mol_names}
\alias{get_mol_names}
\title{Return a character array of names that may be used with the
\code{get_mol_paras} funtion.}
\usage{
get_mol_names()
}
\value{
a character array of names.
}
\description{
Return a character array of names that may be used with the
\code{get_mol_paras} funtion.
}
