% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_parameters.R
\name{get_mol_paras}
\alias{get_mol_paras}
\title{Get a \code{mol_parameters} object for a named molecule.}
\usage{
get_mol_paras(name, ...)
}
\arguments{
\item{name}{the name of the molecule.}

\item{...}{aguments to pass to molecule definition function.}
}
\description{
Get a \code{mol_parameters} object for a named molecule.
}
