% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_mrs_affine}
\alias{get_mrs_affine}
\title{Generate an sform affine for nifti generation.}
\usage{
get_mrs_affine(mrs_data, x_pos = 1, y_pos = 1, z_pos = 1)
}
\arguments{
\item{mrs_data}{input data.}

\item{x_pos}{x_position coordinate.}

\item{y_pos}{y_position coordinate.}

\item{z_pos}{z_position coordinate.}
}
\value{
affine matrix.
}
\description{
Generate an sform affine for nifti generation.
}
