% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_mrsi2d_seg}
\alias{get_mrsi2d_seg}
\title{Calculate the partial volume estimates for each voxel in a 2D MRSI dataset.}
\usage{
get_mrsi2d_seg(mrs_data, mri_seg)
}
\arguments{
\item{mrs_data}{2D MRSI data with multiple voxels in the x-y dimension.}

\item{mri_seg}{MRI data with values corresponding to the segmetation class.}
}
\value{
a data frame of partial volume estimates.
}
\description{
Calculate the partial volume estimates for each voxel in a 2D MRSI dataset.
}
