% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_mrsi_voi}
\alias{get_mrsi_voi}
\title{Generate a MRSI VOI from an \code{mrs_data} object.}
\usage{
get_mrsi_voi(mrs_data, target_mri = NULL, map = NULL, ker = mmand::boxKernel())
}
\arguments{
\item{mrs_data}{MRS data.}

\item{target_mri}{optional image data to match the intended volume space.}

\item{map}{optional voi intensity map.}

\item{ker}{kernel to rescale the map data to the target_mri.}
}
\value{
volume data as a nifti object.
}
\description{
Generate a MRSI VOI from an \code{mrs_data} object.
}
