% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_mrsi_voxel}
\alias{get_mrsi_voxel}
\title{Generate a MRSI voxel from an \code{mrs_data} object.}
\usage{
get_mrsi_voxel(mrs_data, target_mri, x_pos, y_pos, z_pos)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{target_mri}{optional image data to match the intended volume space.}

\item{x_pos}{x voxel coordinate.}

\item{y_pos}{y voxel coordinate.}

\item{z_pos}{z voxel coordinate.}
}
\value{
volume data as a nifti object.
}
\description{
Generate a MRSI voxel from an \code{mrs_data} object.
}
