% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_odd_dyns}
\alias{get_odd_dyns}
\title{Return odd numbered dynamic scans starting from 1 (1,3,5...).}
\usage{
get_odd_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}
}
\value{
dynamic MRS data containing odd numbered scans.
}
\description{
Return odd numbered dynamic scans starting from 1 (1,3,5...).
}
