% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_ref}
\alias{get_ref}
\title{Extract the reference component from an mrs_data object.}
\usage{
get_ref(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
Reference component.
}
\description{
Extract the reference component from an mrs_data object.
}
