% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_slice}
\alias{get_slice}
\title{Return a single slice from a larger MRSI dataset.}
\usage{
get_slice(mrs_data, z_pos)
}
\arguments{
\item{mrs_data}{MRSI data.}

\item{z_pos}{the z index to extract.}
}
\value{
MRS data.
}
\description{
Return a single slice from a larger MRSI dataset.
}
