% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{get_spant_resources_dir}
\alias{get_spant_resources_dir}
\title{Return the spant resources directory.}
\usage{
get_spant_resources_dir()
}
\value{
full path to the spant resources directory.
}
\description{
Usually used to store additional software, datasets and templates. An attempt
will be made to create the folder if not found. The default resources
directory is "spant_resources" located in the users HOME directory.
}
