% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{get_spin_num}
\alias{get_spin_num}
\title{Return the spin number for a given nucleus.}
\usage{
get_spin_num(nucleus)
}
\arguments{
\item{nucleus}{nucleus name, eg "1H".}
}
\value{
spin number.
}
\description{
Return the spin number for a given nucleus.
}
