% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_subset}
\alias{get_subset}
\title{Extract a subset}
\usage{
get_subset(mrs_data, x_set = NULL, y_set = NULL, z_set = NULL,
  dyn_set = NULL, coil_set = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
MRS data.
}
\description{
Extract a subset
}
