% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_svs_voi}
\alias{get_svs_voi}
\title{Generate a SVS acquisition volume from an \code{mrs_data} object.}
\usage{
get_svs_voi(mrs_data, target_mri)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{target_mri}{optional image data to match the intended volume space.}
}
\value{
volume data as a nifti object.
}
\description{
Generate a SVS acquisition volume from an \code{mrs_data} object.
}
