% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_tail_dyns}
\alias{get_tail_dyns}
\title{Return the last scans of a dynamic series.}
\usage{
get_tail_dyns(mrs_data, n = 1)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{n}{the number of dynamic scans to return.}
}
\value{
last scans of a dynamic series.
}
\description{
Return the last scans of a dynamic series.
}
