% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mol_parameters.R
\name{get_uncoupled_mol}
\alias{get_uncoupled_mol}
\title{Generate a \code{mol_parameters} object for a simple spin system with one resonance.}
\usage{
get_uncoupled_mol(name, chem_shift, nucleus, scale_factor, lw, lg)
}
\arguments{
\item{name}{Name of the molecule.}

\item{chem_shift}{Chemical shift of the resonance (PPM).}

\item{nucleus}{Nucleus (1H, 31P...).}

\item{scale_factor}{Multiplicative scaling factor.}

\item{lw}{Line width in Hz.}

\item{lg}{Lorentz-Gauss lineshape parameter (between 0 and 1).}
}
\value{
A mol_parameters object.
}
\description{
Generate a \code{mol_parameters} object for a simple spin system with one resonance.
}
