% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_voi_cog}
\alias{get_voi_cog}
\title{Calculate the centre of gravity for an image containing 0 and 1's.}
\usage{
get_voi_cog(voi)
}
\arguments{
\item{voi}{nifti object.}
}
\value{
triplet of x,y,z coordinates.
}
\description{
Calculate the centre of gravity for an image containing 0 and 1's.
}
