% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{get_voi_seg}
\alias{get_voi_seg}
\title{Return the white matter, gray matter and CSF composition of a volume.}
\usage{
get_voi_seg(voi, mri_seg)
}
\arguments{
\item{voi}{volume data as a nifti object}

\item{mri_seg}{segmented brain volume as a nifti object}
}
\value{
a vector of partial volumes expressed as percentages
}
\description{
Return the white matter, gray matter and CSF composition of a volume.
}
