% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_voxel}
\alias{get_voxel}
\title{Return a single voxel from a larger mrs dataset.}
\usage{
get_voxel(mrs_data, x_pos = 1, y_pos = 1, z_pos = 1, dyn = 1,
  coil = 1)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{x_pos}{the x index to plot.}

\item{y_pos}{the y index to plot.}

\item{z_pos}{the z index to plot.}

\item{dyn}{the dynamic index to plot.}

\item{coil}{the coil element number to plot.}
}
\value{
MRS data.
}
\description{
Return a single voxel from a larger mrs dataset.
}
