% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{glm_spec}
\alias{glm_spec}
\title{Perform a GLM analysis of dynamic MRS data in the spectral domain.}
\usage{
glm_spec(mrs_data, regressor_df)
}
\arguments{
\item{mrs_data}{single-voxel dynamics MRS data.}

\item{regressor_df}{a data frame containing temporal regressors to be applied
to each spectral datapoint.}
}
\value{
list of statistical results.
}
\description{
Perform a GLM analysis of dynamic MRS data in the spectral domain.
}
