% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{glm_spec_group_linhyp}
\alias{glm_spec_group_linhyp}
\title{Test a group-level spectral GLM linear hypothesis.}
\usage{
glm_spec_group_linhyp(hmat, analysis_dir = "spant_analysis")
}
\arguments{
\item{hmat}{linear hypothesis matrix.}

\item{analysis_dir}{directory containing preprocessed data generated by
the preproc_svs_dataset function.}
}
\description{
Test a group-level spectral GLM linear hypothesis.
}
