% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{grid_shift_xy}
\alias{grid_shift_xy}
\title{Grid shift MRSI data in the x/y dimension.}
\usage{
grid_shift_xy(mrs_data, x_shift, y_shift)
}
\arguments{
\item{mrs_data}{MRSI data in the spatial domain.}

\item{x_shift}{shift to apply in the x-direction in units of voxels.}

\item{y_shift}{shift to apply in the y-direction in units of voxels.}
}
\value{
shifted data.
}
\description{
Grid shift MRSI data in the x/y dimension.
}
