% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{gridplot.mrs_data}
\alias{gridplot.mrs_data}
\title{Arrange spectral plots in a grid.}
\usage{
\method{gridplot}{mrs_data}(
  x,
  rows = NA,
  cols = NA,
  mar = c(0, 0, 0, 0),
  oma = c(3.5, 1, 1, 1),
  bty = "o",
  restore_def_par = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{rows}{number of grid rows.}

\item{cols}{number of grid columns.}

\item{mar}{option to adjust the plot margins. See ?par.}

\item{oma}{outer margin area.}

\item{bty}{option to draw a box around the plot. See ?par.}

\item{restore_def_par}{restore default plotting par values after the plot has
been made.}

\item{...}{other arguments to pass to the plot method.}
}
\description{
Arrange spectral plots in a grid.
}
