% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{gridplot}
\alias{gridplot}
\title{Arrange spectral plots in a grid.}
\usage{
gridplot(x, ...)
}
\arguments{
\item{x}{object for plotting.}

\item{...}{arguments to be passed to methods.}
}
\description{
Arrange spectral plots in a grid.
}
