% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{hz}
\alias{hz}
\title{Return the frequency scale of an MRS dataset in Hz.}
\usage{
hz(mrs_data, fs = NULL, N = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{fs}{sampling frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}
}
\value{
frequency scale.
}
\description{
Return the frequency scale of an MRS dataset in Hz.
}
