% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ift_shift}
\alias{ift_shift}
\title{Perform an iffshift and ifft on a vector.}
\usage{
ift_shift(vec_in)
}
\arguments{
\item{vec_in}{vector input.}
}
\value{
output vector.
}
\description{
Perform an iffshift and ifft on a vector.
}
