% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ift_shift_mat}
\alias{ift_shift_mat}
\title{Perform an ifft and ifftshift on a matrix with each column replaced by its
shifted ifft.}
\usage{
ift_shift_mat(mat_in)
}
\arguments{
\item{mat_in}{matrix input.}
}
\value{
output matrix.
}
\description{
Perform an ifft and ifftshift on a matrix with each column replaced by its
shifted ifft.
}
