% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Im.mrs_data}
\alias{Im.mrs_data}
\title{Apply Im operator to an MRS dataset.}
\usage{
\method{Im}{mrs_data}(x)
}
\arguments{
\item{x}{MRS data.}
}
\value{
MRS data following Im operator.
}
\description{
Apply Im operator to an MRS dataset.
}
