% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Imzap}
\alias{Imzap}
\title{Complex rounding function taken from complexplus package to reduce the number
of spant dependencies.}
\usage{
Imzap(x, tol = 1e-06)
}
\arguments{
\item{x}{a scalar or vector, real or complex.}

\item{tol}{a tolerance, 10^-6 by default. Prevents possible numerical
problems. Can be set to 0 if desired.}
}
\description{
Complex rounding function taken from complexplus package to reduce the number
of spant dependencies.
}
