% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{install_faceoff}
\alias{install_faceoff}
\title{Install FaceOff from : https://github.com/srikash/FaceOff, requires ANTs to
be installed to work.}
\usage{
install_faceoff()
}
\description{
Install FaceOff from : https://github.com/srikash/FaceOff, requires ANTs to
be installed to work.
}
