% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{install_oasis_template}
\alias{install_oasis_template}
\title{Install the Oasis brain template from :
https://doi.org/10.6084/m9.figshare.915436.v2}
\usage{
install_oasis_template()
}
\description{
Install the Oasis brain template from :
https://doi.org/10.6084/m9.figshare.915436.v2
}
