% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{int_spec}
\alias{int_spec}
\title{Integrate a spectral region.}
\usage{
int_spec(mrs_data, xlim = NULL, freq_scale = "ppm", mode = "re")
}
\arguments{
\item{mrs_data}{MRS data.}

\item{xlim}{spectral range to be integrated (defaults to full range).}

\item{freq_scale}{units of xlim, can be : "ppm", "hz" or "points".}

\item{mode}{spectral mode, can be : "re", "im", "mod" or "cplx".}
}
\value{
an array of integral values.
}
\description{
See spec_op function for a more complete set of spectral operations.
}
