% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{inv_odd_dyns}
\alias{inv_odd_dyns}
\title{Invert odd numbered dynamic scans starting from 1 (1,3,5...).}
\usage{
inv_odd_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}
}
\value{
Dynamic MRS data with inverted odd numbered scans.
}
\description{
Invert odd numbered dynamic scans starting from 1 (1,3,5...).
}
