% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.def}
\alias{is.def}
\title{Check if an object is defined, which is the same as being not NULL.}
\usage{
is.def(x)
}
\arguments{
\item{x}{object to test for being NULL.}
}
\value{
logical value.
}
\description{
Check if an object is defined, which is the same as being not NULL.
}
