% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{is_fd}
\alias{is_fd}
\title{Check if the chemical shift dimension of an MRS data object is in the
frequency domain.}
\usage{
is_fd(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
logical value.
}
\description{
Check if the chemical shift dimension of an MRS data object is in the
frequency domain.
}
