% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{lb.basis_set}
\alias{lb.basis_set}
\title{Apply line-broadening (apodisation) to a basis-set object.}
\usage{
\method{lb}{basis_set}(basis_set, lb, lg = 1)
}
\arguments{
\item{basis_set}{input basis-set}

\item{lb}{amount of line-broadening in Hz}

\item{lg}{Lorentz-Gauss lineshape parameter (between 0 and 1)}
}
\value{
line-broadened data
}
\description{
Apply line-broadening (apodisation) to a basis-set object.
}
