% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{lb}
\alias{lb}
\title{Apply line-broadening (apodisation) to MRS data.}
\usage{
lb(mrs_data, lb, lg = 1)
}
\arguments{
\item{mrs_data}{input data}

\item{lb}{amount on line-broadening in Hz}

\item{lg}{Lorentz-Gauss lineshape parameter (between 0 and 1)}
}
\value{
line-broadened data
}
\description{
Apply line-broadening (apodisation) to MRS data.
}
