% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{lb_degrade_snr}
\alias{lb_degrade_snr}
\title{Apply line-broadening to dynamic MRS data and degrade the SNR to reverse any
improvement.}
\usage{
lb_degrade_snr(mrs_data, lb, lg = NULL)
}
\arguments{
\item{mrs_data}{data to be broadened.}

\item{lb}{amount of line-broadening in Hz.}

\item{lg}{Lorentz-Gauss lineshape parameter (between 0 and 1). Defaults to 0.}
}
\value{
line-broadened data.
}
\description{
Apply line-broadening to dynamic MRS data and degrade the SNR to reverse any
improvement.
}
