% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lofdc.R
\name{lofdc}
\alias{lofdc}
\title{Correct linear frequency drift.}
\usage{
lofdc(
  mrs_data,
  max_hz_s = 0.1,
  tr = NULL,
  ret_corr_only = TRUE,
  outlier_thresh = 3,
  xlim = c(4, 0.5),
  order = 1
)
}
\arguments{
\item{mrs_data}{MRS data to be corrected.}

\item{max_hz_s}{the maximum drift rate to search over.}

\item{tr}{mrs_data repetition time.}

\item{ret_corr_only}{return the corrected mrs_data object only.}

\item{outlier_thresh}{threshold to remove outliers.}

\item{xlim}{spectral width (in ppm) to evaluate outliers.}

\item{order}{correction order.}
}
\value{
drift corrected mrs_data object.
}
\description{
Correct linear frequency drift.
}
