% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lw2alpha}
\alias{lw2alpha}
\title{Covert a linewidth in Hz to an equivalent alpha value in the time-domain ie:
x * exp(-t * alpha).}
\usage{
lw2alpha(lw)
}
\arguments{
\item{lw}{linewidth in Hz.}
}
\value{
beta damping value.
}
\description{
Covert a linewidth in Hz to an equivalent alpha value in the time-domain ie:
x * exp(-t * alpha).
}
