% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lw2beta}
\alias{lw2beta}
\title{Covert a linewidth in Hz to an equivalent beta value in the time-domain ie:
x * exp(-t * t * beta).}
\usage{
lw2beta(lw)
}
\arguments{
\item{lw}{linewidth in Hz.}
}
\value{
beta damping value.
}
\description{
Covert a linewidth in Hz to an equivalent beta value in the time-domain ie:
x * exp(-t * t * beta).
}
