% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mask_dyns}
\alias{mask_dyns}
\title{Mask an MRS dataset in the dynamic dimension.}
\usage{
mask_dyns(mrs_data, mask)
}
\arguments{
\item{mrs_data}{MRS data object.}

\item{mask}{vector of boolean values specifying the dynamics to mask, where a
value of TRUE indicates the spectrum should be removed.}
}
\value{
masked dataset.
}
\description{
Mask an MRS dataset in the dynamic dimension.
}
