% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mask_fit_res}
\alias{mask_fit_res}
\title{Mask fit result spectra depending on a vector of bool values.}
\usage{
mask_fit_res(fit_result, mask_vec, amps_only = FALSE)
}
\arguments{
\item{fit_result}{fit result object to be masked.}

\item{mask_vec}{a Boolean vector with the same number of rows as there are
rows in the results table.}

\item{amps_only}{only mask the amplitude and associated error estimate
columns.}
}
\value{
a masked fit result object.
}
\description{
Mask fit result spectra depending on a vector of bool values.
}
