% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mask_xy_corners}
\alias{mask_xy_corners}
\title{Mask the four corners of an MRSI dataset in the x-y plane.}
\usage{
mask_xy_corners(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
masked MRS data.
}
\description{
Mask the four corners of an MRSI dataset in the x-y plane.
}
