% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mask_xy_ellipse}
\alias{mask_xy_ellipse}
\title{Mask the voxels outside an elliptical region spanning the MRSI dataset in the
x-y plane.}
\usage{
mask_xy_ellipse(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
masked MRS data.
}
\description{
Mask the voxels outside an elliptical region spanning the MRSI dataset in the
x-y plane.
}
