% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mat2mrs_data}
\alias{mat2mrs_data}
\title{Convert a matrix (with dynamics in the row dimension and spectral points in
the column dimensions) into a mrs_data object.}
\usage{
mat2mrs_data(mat, fs = def_fs(), ft = def_ft(), ref = def_ref(),
  fd = FALSE)
}
\arguments{
\item{mat}{The data matrix.}

\item{fs}{Sampling frequency in Hz.}

\item{ft}{Transmitter frequency in Hz.}

\item{ref}{Reference value for ppm scale.}

\item{fd}{Flag to indicate if the matrix is in the frequency domain (logical).}
}
\description{
Convert a matrix (with dynamics in the row dimension and spectral points in
the column dimensions) into a mrs_data object.
}
