% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_files}
\alias{match_files}
\title{Match files based on a vector of input paths and a glob pattern. The glob
pattern is appended to each path and should match one file only.}
\usage{
match_files(paths, glob)
}
\arguments{
\item{paths}{vectors of filesystem paths.}

\item{glob}{pattern to append to each path before passing to Sys.glob.}
}
\value{
matched files. NA values correspond to either no match or multiple
matches.
}
\description{
Match files based on a vector of input paths and a glob pattern. The glob
pattern is appended to each path and should match one file only.
}
