% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matexp}
\alias{matexp}
\title{Matrix exponential function taken from complexplus package to reduce the
number of spant dependencies.}
\usage{
matexp(x)
}
\arguments{
\item{x}{a square complex matrix.}
}
\value{
the matrix exponential of x.
}
\description{
Matrix exponential function taken from complexplus package to reduce the
number of spant dependencies.
}
