% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{max.mrs_data}
\alias{max.mrs_data}
\title{Apply the max operator to an MRS dataset.}
\usage{
\method{max}{mrs_data}(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
MRS data following max operator.
}
\description{
Apply the max operator to an MRS dataset.
}
