% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{max_mrs}
\alias{max_mrs}
\title{Apply the max operator to an MRS dataset.}
\usage{
max_mrs(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
MRS data following max operator.
}
\description{
Apply the max operator to an MRS dataset.
}
