% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{max_mrs_interp}
\alias{max_mrs_interp}
\title{Apply the max operator to an interpolated MRS dataset.}
\usage{
max_mrs_interp(mrs_data, interp_f = 4)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{interp_f}{interpolation factor.}
}
\value{
Array of maximum values (real only).
}
\description{
Apply the max operator to an interpolated MRS dataset.
}
