% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mean.list}
\alias{mean.list}
\title{Calculate the mean spectrum from an mrs_data object.}
\usage{
\method{mean}{list}(x, ...)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{...}{other arguments to pass to the colMeans function.}
}
\value{
mean mrs_data object.
}
\description{
Calculate the mean spectrum from an mrs_data object.
}
