% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mean_dyn_pairs}
\alias{mean_dyn_pairs}
\title{Calculate the pairwise means across a dynamic data set.}
\usage{
mean_dyn_pairs(mrs_data)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}
}
\value{
mean dynamic data of adjacent dynamic pairs.
}
\description{
Calculate the pairwise means across a dynamic data set.
}
