% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mean_dyns}
\alias{mean_dyns}
\title{Calculate the mean dynamic data.}
\usage{
mean_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}
}
\value{
Mean dynamic data.
}
\description{
Calculate the mean dynamic data.
}
