% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mean_mrs_list}
\alias{mean_mrs_list}
\title{Return the mean of a list of mrs_data objects.}
\usage{
mean_mrs_list(mrs_list)
}
\arguments{
\item{mrs_list}{list of mrs_data objects.}
}
\value{
mean \code{mrs_data} object.
}
\description{
Return the mean of a list of mrs_data objects.
}
