% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_vec_blocks}
\alias{mean_vec_blocks}
\title{Calculate the mean of adjacent blocks in a vector.}
\usage{
mean_vec_blocks(x, block_size)
}
\arguments{
\item{x}{input vector.}

\item{block_size}{number of adjacent elements to average over.}
}
\value{
vector data averaged in blocks.
}
\description{
Calculate the mean of adjacent blocks in a vector.
}
