% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{median_dyns}
\alias{median_dyns}
\title{Calculate the median dynamic data.}
\usage{
median_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}
}
\value{
Median dynamic data.
}
\description{
Calculate the median dynamic data.
}
