% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Mod.mrs_data}
\alias{Mod.mrs_data}
\title{Apply Mod operator to an MRS dataset.}
\usage{
\method{Mod}{mrs_data}(z)
}
\arguments{
\item{z}{MRS data.}
}
\value{
MRS data following Mod operator.
}
\description{
Apply Mod operator to an MRS dataset.
}
