% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mod_td}
\alias{mod_td}
\title{Apply the Modulus operator to the time-domain MRS signal.}
\usage{
mod_td(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data input.}
}
\value{
time-domain modulus of input.
}
\description{
Apply the Modulus operator to the time-domain MRS signal.
}
