% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{mrs_data2basis}
\alias{mrs_data2basis}
\title{Convert an mrs_data object to basis object - where basis signals are spread
across the dynamic dimension in the MRS data.}
\usage{
mrs_data2basis(mrs_data, names)
}
\arguments{
\item{mrs_data}{mrs_data object with basis signals spread across the dynamic dimension.}

\item{names}{list of names corresponding to basis signals.}
}
\value{
basis set object.
}
\description{
Convert an mrs_data object to basis object - where basis signals are spread
across the dynamic dimension in the MRS data.
}
