% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mrs_data2list}
\alias{mrs_data2list}
\title{Split MRS data containing multiple spectra into a list of single spectra
datasets.}
\usage{
mrs_data2list(mrs_data)
}
\arguments{
\item{mrs_data}{input MRS dataset}
}
\value{
list of MRS datasets
}
\description{
Split MRS data containing multiple spectra into a list of single spectra
datasets.
}
