% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mrs_data2mat}
\alias{mrs_data2mat}
\title{Convert mrs_data object to a matrix, with dynamics in the row dimension
and spectal points in the column dimension.}
\usage{
mrs_data2mat(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data object.}
}
\value{
A matrix.
}
\description{
Convert mrs_data object to a matrix, with dynamics in the row dimension
and spectal points in the column dimension.
}
