% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mrs_data2vec}
\alias{mrs_data2vec}
\title{Convert mrs_data object to a vector.}
\usage{
mrs_data2vec(mrs_data, dyn = 1, x_pos = 1, y_pos = 1, z_pos = 1,
  coil = 1)
}
\arguments{
\item{dyn}{dynamic index.}

\item{x_pos}{x index.}

\item{y_pos}{y index.}

\item{z_pos}{z index.}

\item{coil}{coil element index.}
}
\value{
MRS data vector.
}
\description{
Convert mrs_data object to a vector.
}
