% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{mrs_data_list2bids}
\alias{mrs_data_list2bids}
\title{Create a BIDS directory and file structure from a list of mrs_data objects.}
\usage{
mrs_data_list2bids(
  mrs_data_list,
  output_dir,
  runs = 1,
  sessions = 1,
  sub_labels = NULL
)
}
\arguments{
\item{mrs_data_list}{list of mrs_data objects.}

\item{output_dir}{the base directory to create the BIDS structure.}

\item{runs}{number of runs per subject and session.}

\item{sessions}{number of sessions.}

\item{sub_labels}{optional labels for subject level identification.}
}
\description{
Create a BIDS directory and file structure from a list of mrs_data objects.
}
