% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mrsi2d_img2kspace}
\alias{mrsi2d_img2kspace}
\title{Transform 2D MRSI data to k-space in the x-y direction.}
\usage{
mrsi2d_img2kspace(mrs_data)
}
\arguments{
\item{mrs_data}{2D MRSI data.}
}
\value{
k-space data.
}
\description{
Transform 2D MRSI data to k-space in the x-y direction.
}
