% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mrsi2d_kspace2img}
\alias{mrsi2d_kspace2img}
\title{Transform 2D MRSI data from k-space to image space in the x-y direction.}
\usage{
mrsi2d_kspace2img(mrs_data)
}
\arguments{
\item{mrs_data}{2D MRSI data.}
}
\value{
MRSI data in image space.
}
\description{
Transform 2D MRSI data from k-space to image space in the x-y direction.
}
