% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvfftshift}
\alias{mvfftshift}
\title{Perform a fftshift on a matrix, with each column replaced by its shifted
result.}
\usage{
mvfftshift(x)
}
\arguments{
\item{x}{matrix input.}
}
\value{
output matrix.
}
\description{
Perform a fftshift on a matrix, with each column replaced by its shifted
result.
}
