% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvifftshift}
\alias{mvifftshift}
\title{Perform an ifftshift on a matrix, with each column replaced by its shifted
result.}
\usage{
mvifftshift(x)
}
\arguments{
\item{x}{matrix input.}
}
\value{
output matrix.
}
\description{
Perform an ifftshift on a matrix, with each column replaced by its shifted
result.
}
