% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{n2coord}
\alias{n2coord}
\title{Print fit coordinates from a single index.}
\usage{
n2coord(n, fit_res)
}
\arguments{
\item{n}{fit index.}

\item{fit_res}{\code{fit_result} object.}
}
\description{
Print fit coordinates from a single index.
}
