% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Ncoils}
\alias{Ncoils}
\title{Return the total number of coil elements in an MRS dataset.}
\usage{
Ncoils(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of coil elements in an MRS dataset.
}
