% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{nifti_flip_lr}
\alias{nifti_flip_lr}
\title{Flip the x data dimension order of a nifti image. This corresponds to
flipping MRI data in the left-right direction, assuming the data in save in
neurological format (can check with fslorient program).}
\usage{
nifti_flip_lr(x)
}
\arguments{
\item{x}{nifti object to be processed.}
}
\value{
nifti object with reversed x data direction.
}
\description{
Flip the x data dimension order of a nifti image. This corresponds to
flipping MRI data in the left-right direction, assuming the data in save in
neurological format (can check with fslorient program).
}
