% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Npts}
\alias{Npts}
\title{Return the number of data points in an MRS dataset.}
\usage{
Npts(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
number of data points.
}
\description{
Return the number of data points in an MRS dataset.
}
