% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Nspec}
\alias{Nspec}
\title{Return the total number of spectra in an MRS dataset.}
\usage{
Nspec(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of spectra in an MRS dataset.
}
