% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Ntrans}
\alias{Ntrans}
\title{Return the total number of acquired transients for an MRS dataset.}
\usage{
Ntrans(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of acquired transients for an MRS dataset.
}
