% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Nx}
\alias{Nx}
\title{Return the total number of x locations in an MRS dataset.}
\usage{
Nx(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of x locations in an MRS dataset.
}
