% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Ny}
\alias{Ny}
\title{Return the total number of y locations in an MRS dataset.}
\usage{
Ny(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of y locations in an MRS dataset.
}
