% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Nz}
\alias{Nz}
\title{Return the total number of z locations in an MRS dataset.}
\usage{
Nz(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\description{
Return the total number of z locations in an MRS dataset.
}
