% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{one_page_pdf}
\alias{one_page_pdf}
\title{Export a one-page pdf of a single fit result}
\usage{
one_page_pdf(fit_res, path)
}
\arguments{
\item{fit_res}{\code{fit_result} object.}

\item{path}{path to the exported pdf file.}
}
\description{
Export a one-page pdf of a single fit result
}
