% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{ortho3}
\alias{ortho3}
\title{Display an orthographic projection plot of a nifti object.}
\usage{
ortho3(
  underlay,
  overlay = NULL,
  xyz = NULL,
  zlim = NULL,
  zlim_ol = NULL,
  alpha = 1,
  col_ol = viridisLite::viridis(64),
  orient_lab = TRUE,
  rescale = 1,
  crosshairs = TRUE
)
}
\arguments{
\item{underlay}{underlay image to be shown in grayscale.}

\item{overlay}{optional overlay image.}

\item{xyz}{x, y, z slice coordinates to display.}

\item{zlim}{underlay intensity limits.}

\item{zlim_ol}{overlay intensity limits.}

\item{alpha}{transparency of overlay.}

\item{col_ol}{color palette of overlay.}

\item{orient_lab}{display orientation labels (default TRUE).}

\item{rescale}{rescale factor for the underlay and overlay images.}

\item{crosshairs}{display the crosshairs (default TRUE).}
}
\description{
Display an orthographic projection plot of a nifti object.
}
