% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{ortho3_int}
\alias{ortho3_int}
\title{Display an interactive orthographic projection plot of a nifti object.}
\usage{
ortho3_int(
  underlay,
  overlay = NULL,
  xyz = NULL,
  zlim = NULL,
  zlim_ol = NULL,
  alpha = 1,
  ...
)
}
\arguments{
\item{underlay}{underlay image to be shown in grayscale.}

\item{overlay}{optional overlay image.}

\item{xyz}{x, y, z slice coordinates to display.}

\item{zlim}{underlay intensity limits.}

\item{zlim_ol}{overlay intensity limits.}

\item{alpha}{transparency of overlay.}

\item{...}{other options to be passed to the ortho3 function.}
}
\description{
Display an interactive orthographic projection plot of a nifti object.
}
