% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paths2df}
\alias{paths2df}
\title{Split paths into parts based on backslash, forwardslash and dot characters
and return a data frame of these parts.}
\usage{
paths2df(paths, col_num = NULL, extra_regex = NULL)
}
\arguments{
\item{paths}{a vector of paths.}

\item{col_num}{when set, only the specified column number of the data frame
will be returned.}

\item{extra_regex}{extra regular expression for splitting the paths. Example,
"_|-" could be used to additionally split on underscores and hyphen
characters.}
}
\value{
data.frame of separated path elements.
}
\description{
Split paths into parts based on backslash, forwardslash and dot characters
and return a data frame of these parts.
}
