% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{plot.fit_result}
\alias{plot.fit_result}
\title{Plot the fitting results of an object of class fit_result.}
\usage{
\method{plot}{fit_result}(x, n = NA, ...)
}
\arguments{
\item{x}{fit_result object.}

\item{n}{index to the fit number to display.}

\item{...}{further arguments.}
}
\description{
Plot the fitting results of an object of class fit_result.
}
