% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{plot.mrs_data}
\alias{plot.mrs_data}
\title{TODO document.}
\usage{
\method{plot}{mrs_data}(x, fd = TRUE, scale = NULL, xlim = NULL,
  yscale = FALSE, mode = "real", dyn = 1, x_pos = 1, y_pos = 1,
  z_pos = 1, coil = 1, ref = 1, lwd = NULL, bty = NULL, label = "",
  ...)
}
\description{
TODO document.
}
