% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{plot_bc}
\alias{plot_bc}
\title{Convenience function to plot a baseline estimate with the original data.}
\usage{
plot_bc(orig_data, bc_data, ...)
}
\arguments{
\item{orig_data}{the original data.}

\item{bc_data}{the baseline corrected data.}

\item{...}{other arguments to pass to the stackplot function.}
}
\description{
Convenience function to plot a baseline estimate with the original data.
}
