% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{plot_slice_fit_inter}
\alias{plot_slice_fit_inter}
\title{Plot a 2D slice from an MRSI fit result object.}
\usage{
plot_slice_fit_inter(
  fit_res,
  map = NULL,
  map_denom = NULL,
  slice = 1,
  zlim = NULL,
  interp = 1,
  xlim = NULL
)
}
\arguments{
\item{fit_res}{\code{fit_result} object.}

\item{map}{fit result values to display as a colour map. Can be specified as
a charactor string or array of numeric values. Defaults to "tNAA".}

\item{map_denom}{fit result values to divide the map argument by. Can be
specified as a charactor string (eg "tCr") or array of numeric values.}

\item{slice}{slice to plot in the z direction.}

\item{zlim}{range of values to plot.}

\item{interp}{interpolation factor.}

\item{xlim}{spectral plot limits for the x axis.}
}
\description{
Plot a 2D slice from an MRSI fit result object.
}
