% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{plot_slice_map}
\alias{plot_slice_map}
\title{Plot a slice from a 7 dimensional array.}
\usage{
plot_slice_map(data, zlim = NULL, mask_map = NULL, mask_cutoff = 20,
  interp = 1, slice = 1, dyn = 1, coil = 1, ref = 1,
  denom = NULL, horizontal = FALSE)
}
\arguments{
\item{data}{7d array of values to be plotted.}

\item{zlim}{smallest and largest values to be plotted.}

\item{mask_map}{matching map with logical values to indicate if the
corresponding values should be plotted.}

\item{mask_cutoff}{minimum values to plot (as a percentage of the maximum).}

\item{interp}{map interpolation factor.}

\item{slice}{the slice index to plot.}

\item{dyn}{the dynamic index to plot.}

\item{coil}{the coil element number to plot.}

\item{ref}{reference index to plot.}

\item{denom}{map to use as a denominator.}

\item{horizontal}{display the colorbar horizontally (logical).}
}
\description{
Plot a slice from a 7 dimensional array.
}
