% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{plot_slice_map_inter}
\alias{plot_slice_map_inter}
\title{Plot an interactive slice map from a data array where voxels can be selected
to display a corresponding spectrum}
\usage{
plot_slice_map_inter(map, mrs_data, xlim = NULL, slice = 1,
  mask_map = NULL, upper = NULL, lower = NULL, denom = NULL,
  mask_cutoff = 20, interp = 1)
}
\arguments{
\item{map}{array of values to be plotted}

\item{mrs_data}{spectral data}

\item{xlim}{spectral region to plot}

\item{slice}{the slice index to plot}

\item{mask_map}{matching map with logical values to indicate if the
corresponding values should be plotted}

\item{upper}{cap on the largest values to be plotted}

\item{lower}{cap on the smallest values to be plotted}

\item{denom}{map to use as a denominator}

\item{mask_cutoff}{minimum values to plot (as a percentage of the maximum)}

\item{interp}{map interpolation factor}
}
\description{
Plot an interactive slice map from a data array where voxels can be selected
to display a corresponding spectrum
}
