% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{plot_spec_sd}
\alias{plot_spec_sd}
\title{Plot the spectral standard deviation.}
\usage{
plot_spec_sd(mrs_data, xlim = NULL, scale_sd = 5, ...)
}
\arguments{
\item{mrs_data}{MRS data to be plotted.}

\item{xlim}{plotting limits in ppm.}

\item{scale_sd}{scaling factor for the standard deviation trace.}

\item{...}{other arguments to pass to the stackplot function.}
}
\description{
Plot the spectral standard deviation.
}
