% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{plot_tqn_fit_csv}
\alias{plot_tqn_fit_csv}
\title{Plot a TARQUIN fit from a csv formatted file.}
\usage{
plot_tqn_fit_csv(file, ...)
}
\arguments{
\item{file}{csv formatted fit file produced by TARQUIN.}

\item{...}{further arguments to plot method.}
}
\description{
Plot a TARQUIN fit from a csv formatted file.
}
