% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{plot_voi_overlay}
\alias{plot_voi_overlay}
\title{Plot a volume as an image overlay.}
\usage{
plot_voi_overlay(mri, voi, export_path = NULL, zlim = NULL, ...)
}
\arguments{
\item{mri}{image data as a nifti object or path to data file.}

\item{voi}{volume data as a nifti object or path to data file.}

\item{export_path}{optional path to save the image in png format.}

\item{zlim}{underlay intensity limits.}

\item{...}{additional arguments to the ortho3 function.}
}
\description{
Plot a volume as an image overlay.
}
