% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{ppm}
\alias{ppm}
\title{Return the ppm scale of an MRS dataset.}
\usage{
ppm(mrs_data, ft = NULL, ref = NULL, fs = NULL, N = NULL)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{fs}{sampline frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}
}
\value{
ppm scale.
}
\description{
Return the ppm scale of an MRS dataset.
}
