% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp.R
\name{precomp}
\alias{precomp}
\title{Save function results to file and load on subsequent calls to avoid repeat
computation.}
\usage{
precomp(file, fun, ...)
}
\arguments{
\item{file}{file name to write the results.}

\item{fun}{function to run.}

\item{...}{arguments to be passed to fun.}
}
\description{
Save function results to file and load on subsequent calls to avoid repeat
computation.
}
