% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{preproc_fmrs}
\alias{preproc_fmrs}
\title{Preprocess and perform quality assessment of SVS fMRS data}
\usage{
preproc_fmrs(path, label = NULL, output_dir = NULL)
}
\arguments{
\item{path}{path to the fMRS data file or IMA directory.}

\item{label}{a label to describe the data file.}

\item{output_dir}{output directory.}
}
\description{
Preprocess and perform quality assessment of SVS fMRS data
}
