% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{preproc_fmrs_dataset}
\alias{preproc_fmrs_dataset}
\title{Preprocess and perform quality assessment of one or more SVS fMRS data sets.}
\usage{
preproc_fmrs_dataset(paths, labels = NULL, output_dir = "fmrs_analysis")
}
\arguments{
\item{paths}{paths to the fMRS data file or IMA directory.}

\item{labels}{labels to describe each data set.}

\item{output_dir}{output directory.}
}
\description{
Preprocess and perform quality assessment of one or more SVS fMRS data sets.
}
