% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{preproc_svs}
\alias{preproc_svs}
\title{Preprocess and perform quality assessment of a single SVS data set.}
\usage{
preproc_svs(path, label = NULL, output_dir = NULL)
}
\arguments{
\item{path}{path to the fMRS data file or IMA directory.}

\item{label}{a label to describe the data set.}

\item{output_dir}{output directory.}
}
\description{
Preprocess and perform quality assessment of a single SVS data set.
}
