% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{preproc_svs_dataset}
\alias{preproc_svs_dataset}
\title{Preprocess and perform quality assessment of one or more SVS data sets.}
\usage{
preproc_svs_dataset(
  paths,
  labels = NULL,
  output_dir = "spant_analysis",
  exclude_labels = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{paths}{paths to the fMRS data file or IMA directory.}

\item{labels}{labels to describe each data set.}

\item{output_dir}{output directory.}

\item{exclude_labels}{vector of labels of scans to exclude, eg poor quality
data.}

\item{overwrite}{overwrite saved results, defaults to FALSE.}
}
\description{
Preprocess and perform quality assessment of one or more SVS data sets.
}
