% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{print.fit_result}
\alias{print.fit_result}
\title{Print a summary of an object of class fit_result.}
\usage{
\method{print}{fit_result}(x, ...)
}
\arguments{
\item{x}{fit_result object.}

\item{...}{further arguments.}
}
\description{
Print a summary of an object of class fit_result.
}
