% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{print.mrs_data}
\alias{print.mrs_data}
\title{Print a summary of mrs_data parameters.}
\usage{
\method{print}{mrs_data}(x, full = FALSE, ...)
}
\arguments{
\item{x}{mrs_data object.}

\item{full}{print all parameters (default FALSE).}

\item{...}{further arguments.}
}
\description{
Print a summary of mrs_data parameters.
}
