% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{pulse_acquire}
\alias{pulse_acquire}
\title{Simple pulse and acquire sequence with ideal pulses.}
\usage{
pulse_acquire(spin_params, ft, ref)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}
}
\value{
a list of resonance amplitudes and frequencies.
}
\description{
Simple pulse and acquire sequence with ideal pulses.
}
