% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{Re.mrs_data}
\alias{Re.mrs_data}
\title{Apply Re operator to an MRS dataset.}
\usage{
\method{Re}{mrs_data}(z)
}
\arguments{
\item{z}{MRS data.}
}
\value{
MRS data following Re operator.
}
\description{
Apply Re operator to an MRS dataset.
}
