% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{re_weighting}
\alias{re_weighting}
\title{Apply a weighting to the FID to enchance spectral resolution.}
\usage{
re_weighting(mrs_data, re, alpha)
}
\arguments{
\item{mrs_data}{data to be enhanced.}

\item{re}{resolution enhancement factor (rising exponential factor).}

\item{alpha}{alpha factor (Guassian decay)}
}
\value{
resolution enhanced mrs_data.
}
\description{
Apply a weighting to the FID to enchance spectral resolution.
}
