% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{read_basis}
\alias{read_basis}
\title{Read a basis file in LCModel .basis format.}
\usage{
read_basis(basis_file, ref = def_ref())
}
\arguments{
\item{basis_file}{Path to basis file.}

\item{ref}{Assumed ppm reference value.}
}
\value{
Basis object.
}
\description{
Read a basis file in LCModel .basis format.
}
