% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{read_basis_niidir}
\alias{read_basis_niidir}
\title{Read a basis folder containing one NIfTI MRS file for each basis element.}
\usage{
read_basis_niidir(basis_dir)
}
\arguments{
\item{basis_dir}{directory of basis files.}
}
\value{
a basis_set object.
}
\description{
Read a basis folder containing one NIfTI MRS file for each basis element.
}
