% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_dkd_moco_log}
\alias{read_dkd_moco_log}
\title{Read the log from Dinesh's MoCo sLASER sequence.}
\usage{
read_dkd_moco_log(path, date, end_time)
}
\arguments{
\item{path}{path to the log file.}

\item{date}{scan date, eg "01-01-2025".}

\item{end_time}{scan time, eg "14:00".}
}
\value{
data.frame of moco parameters.
}
\description{
Read the log from Dinesh's MoCo sLASER sequence.
}
