% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{read_fslmrs_json_basis_dir}
\alias{read_fslmrs_json_basis_dir}
\title{Read a basis folder containing one FSL-MRS json file for each basis element.}
\usage{
read_fslmrs_json_basis_dir(basis_dir)
}
\arguments{
\item{basis_dir}{path to directory containing json files.}
}
\value{
a basis_set object.
}
\description{
Read a basis folder containing one FSL-MRS json file for each basis element.
}
