% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{read_fslmrs_json_basis_file}
\alias{read_fslmrs_json_basis_file}
\title{Read an FSL-MRS json file containing a basis signal.}
\usage{
read_fslmrs_json_basis_file(basis_file)
}
\arguments{
\item{basis_file}{path to basis file.}
}
\value{
basis_set object.
}
\description{
Read an FSL-MRS json file containing a basis signal.
}
