% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_read_ima.R
\name{read_ima_coil_dir}
\alias{read_ima_coil_dir}
\title{Read a directory containing Siemens MRS IMA files and combine along the coil
dimension. Note that the coil ID is inferred from the sorted file name and
should be checked when consistency is required between two directories.}
\usage{
read_ima_coil_dir(dir, extra = NULL)
}
\arguments{
\item{dir}{data directory path.}

\item{extra}{an optional data frame with one row to provide additional
variables to be used in subsequent analysis steps, eg id or grouping
variables.}
}
\value{
mrs_data object.
}
\description{
Read a directory containing Siemens MRS IMA files and combine along the coil
dimension. Note that the coil ID is inferred from the sorted file name and
should be checked when consistency is required between two directories.
}
