% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_io.R
\name{read_mrs_dpt}
\alias{read_mrs_dpt}
\title{Read MRS data stored in dangerplot (dpt) v3 format.}
\usage{
read_mrs_dpt(fname)
}
\arguments{
\item{fname}{The filename of the dpt format MRS data.}
}
\value{
An MRS data object.
}
\description{
Read MRS data stored in dangerplot (dpt) v3 format.
}
\examples{
\dontrun{
mrs_data <- read_mrs_dpt(system.file("extdata","svs.dpt",package="spant"))
}
}
