% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_shapes.R
\name{read_pulse_ascii}
\alias{read_pulse_ascii}
\title{Read an ASCII formatted pulse file.}
\usage{
read_pulse_ascii(fname, deg2rad = TRUE)
}
\arguments{
\item{fname}{ASCII formatted pulse file path.}

\item{deg2rad}{convert phase values stored in degrees to radians.}
}
\value{
pulse waveform and header.
}
\description{
Read an ASCII formatted pulse file.
}
