% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_shapes.R
\name{read_pulse_bruker}
\alias{read_pulse_bruker}
\title{Read a Bruker formatted pulse file}
\usage{
read_pulse_bruker(fname)
}
\arguments{
\item{fname}{Bruker formatted pulse file path.}
}
\value{
pulse waveform and header.
}
\description{
Read a Bruker formatted pulse file
}
