% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_read_twix.R
\name{read_siemens_txt_hdr}
\alias{read_siemens_txt_hdr}
\title{Read the text format header found in Siemens IMA and TWIW data files.}
\usage{
read_siemens_txt_hdr(fname, version = "vd")
}
\arguments{
\item{fname}{file name to read.}

\item{version}{software version, can be "vb" or "vd".}
}
\value{
a list of parameter values
}
\description{
Read the text format header found in Siemens IMA and TWIW data files.
}
