% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{read_tqn_fit}
\alias{read_tqn_fit}
\title{Reader for csv fit results generated by TARQUIN.}
\usage{
read_tqn_fit(fit_f)
}
\arguments{
\item{fit_f}{TARQUIN fit file.}
}
\value{
A data frame of the fit data points.
}
\description{
Reader for csv fit results generated by TARQUIN.
}
\examples{
\dontrun{
fit <- read_tqn_fit(system.file("extdata","fit.csv",package="spant"))
}
}
