% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{read_tqn_result}
\alias{read_tqn_result}
\title{Reader for csv results generated by TARQUIN.}
\usage{
read_tqn_result(result_f, remove_rcs = TRUE)
}
\arguments{
\item{result_f}{TARQIUN result file.}

\item{remove_rcs}{Omit row, column and slice ids from output.}
}
\value{
A list of amplitudies, crlbs and diagnostics.
}
\description{
Reader for csv results generated by TARQUIN.
}
\examples{
\dontrun{
result <- read_tqn_result(system.file("extdata","result.csv",package="spant"))
}
}
