% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{recon_imag}
\alias{recon_imag}
\title{Reconstruct complex time-domain data from the real part of frequency-domain
data.}
\usage{
recon_imag(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
reconstructed MRS data.
}
\description{
Reconstruct complex time-domain data from the real part of frequency-domain
data.
}
