% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{recon_imag_vec}
\alias{recon_imag_vec}
\title{Reconstruct complex time-domain data from the real part of frequency-domain
data.}
\usage{
recon_imag_vec(data)
}
\arguments{
\item{data}{data points in the frequency domain.}
}
\value{
reconstructed signal.
}
\description{
Reconstruct complex time-domain data from the real part of frequency-domain
data.
}
