% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{recon_twix_2d_mrsi}
\alias{recon_twix_2d_mrsi}
\title{Reconstruct 2D MRSI data from a twix file loaded with read_mrs.}
\usage{
recon_twix_2d_mrsi(twix_mrs)
}
\arguments{
\item{twix_mrs}{raw dynamic data.}
}
\value{
reconstructed data.
}
\description{
Reconstruct 2D MRSI data from a twix file loaded with read_mrs.
}
