% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rectangular_mask}
\alias{rectangular_mask}
\title{Create a rectangular mask stored as a matrix of logical values.}
\usage{
rectangular_mask(xN, yN, x0, y0, xw, yw, angle)
}
\arguments{
\item{xN}{number of pixels in the x dimension.}

\item{yN}{number of pixels in the y dimension.}

\item{x0}{centre of rectangle in the x direction in units of pixels.}

\item{y0}{centre of rectangle in the y direction in units of pixels.}

\item{xw}{width in the x direction in units of pixels.}

\item{yw}{width in the y direction in units of pixels.}

\item{angle}{angle of rotation in degrees.}
}
\value{
logical mask matrix with dimensions fov_yN x fov_xN.
}
\description{
Create a rectangular mask stored as a matrix of logical values.
}
