% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rep_array_dim}
\alias{rep_array_dim}
\title{Repeat an array over a given dimension.}
\usage{
rep_array_dim(x, rep_dim, n)
}
\arguments{
\item{x}{array.}

\item{rep_dim}{dimension to extend.}

\item{n}{number of times to repeat.}
}
\value{
extended array.
}
\description{
Repeat an array over a given dimension.
}
