% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{rep_dyn}
\alias{rep_dyn}
\title{Replicate a scan in the dynamic dimension.}
\usage{
rep_dyn(mrs_data, times)
}
\arguments{
\item{mrs_data}{MRS data to be replicated.}

\item{times}{Number of times to replicate.}
}
\value{
Replicated data object.
}
\description{
Replicate a scan in the dynamic dimension.
}
