% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{resample_basis}
\alias{resample_basis}
\title{Resample a basis-set to match a mrs_data acquisition}
\usage{
resample_basis(basis, mrs_data, ref_freq_match = TRUE)
}
\arguments{
\item{basis}{the basis to be resampled.}

\item{mrs_data}{the mrs_data to match the number of data points and sampling
frequency.}

\item{ref_freq_match}{apply a frequency shift to the basis to match the
reference frequency (usually 4.65 or 4.68) of the mrs_data.}
}
\value{
resampled basis set object.
}
\description{
Resample a basis-set to match a mrs_data acquisition
}
