% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{resample_voi}
\alias{resample_voi}
\title{Resample a VOI to match a target image space.}
\usage{
resample_voi(voi, mri)
}
\arguments{
\item{voi}{volume data as a nifti object.}

\item{mri}{image data as a nifti object.}
}
\value{
volume data as a nifti object.
}
\description{
Resample a VOI to match a target image space.
}
