% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{reslice_to_mrs}
\alias{reslice_to_mrs}
\title{Reslice a nifti object to match the orientation of mrs data.}
\usage{
reslice_to_mrs(mri, mrs, interp = 3L)
}
\arguments{
\item{mri}{nifti object to be resliced.}

\item{mrs}{mrs_data object for the target orientation.}

\item{interp}{interpolation parameter, see nifyreg.linear definition.}
}
\value{
resliced imaging data.
}
\description{
Reslice a nifti object to match the orientation of mrs data.
}
