% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{rm_basis_elements}
\alias{rm_basis_elements}
\title{Remove elements from a basis set object.}
\usage{
rm_basis_elements(basis, rm_str)
}
\arguments{
\item{basis}{input basis.}

\item{rm_str}{a grep expression to remove basis elements. Use "|" for
multiple matches, eg to match alanine and lactate only : "^Ala$|^Lac$".}
}
\value{
basis with elements removed.
}
\description{
Remove elements from a basis set object.
}
