% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{rm_dyns}
\alias{rm_dyns}
\title{Remove a subset of dynamic scans.}
\usage{
rm_dyns(mrs_data, subset)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{subset}{vector containing indices to the dynamic scans to be
removed.}
}
\value{
MRS data without the specified dynamic scans.
}
\description{
Remove a subset of dynamic scans.
}
