% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_scaling.R
\name{scale_amp_ratio_value}
\alias{scale_amp_ratio_value}
\title{Scale fitted amplitudes to a ratio of signal amplitude.}
\usage{
scale_amp_ratio_value(fit_result, value)
}
\arguments{
\item{fit_result}{a result object generated from fitting.}

\item{value}{the number use as a denominator.}
}
\value{
a \code{fit_result} object with a rescaled results table.
}
\description{
Scale fitted amplitudes to a ratio of signal amplitude.
}
