% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{scale_basis_amp}
\alias{scale_basis_amp}
\title{Scale a basis object by a scalar.}
\usage{
scale_basis_amp(basis, amp)
}
\arguments{
\item{basis}{basis_set object to be scaled.}

\item{amp}{multiplicative factor with length 1.}
}
\value{
basis_set object multiplied by the amplitude scale factor.
}
\description{
Scale a basis object by a scalar.
}
