% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{scale_mrs}
\alias{scale_mrs}
\title{Scale an mrs_data object by a constant.}
\usage{
scale_mrs(mrs_data, scale)
}
\arguments{
\item{mrs_data}{data to be scaled.}

\item{scale}{multiplicative factor.}
}
\value{
mrs_data multiplied by the scale factor.
}
\description{
Scale an mrs_data object by a constant.
}
