% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sd.mrs_data}
\alias{sd.mrs_data}
\title{Calculate the standard deviation spectrum from an mrs_data object.}
\usage{
\method{sd}{mrs_data}(x, na.rm = FALSE)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{na.rm}{remove NA values.}
}
\value{
sd mrs_data object.
}
\description{
Calculate the standard deviation spectrum from an mrs_data object.
}
