% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{seconds}
\alias{seconds}
\title{Return a time scale vector to match the FID of an MRS data object.}
\usage{
seconds(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
A time scale vector in units of seconds.
}
\description{
Return a time scale vector to match the FID of an MRS data object.
}
