% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{segment_t1_rpyants}
\alias{segment_t1_rpyants}
\title{Segment T1 weighted MRI data using the rpyANTs interface to ANTs and write to
file.}
\usage{
segment_t1_rpyants(mri_path, out_dir = NULL)
}
\arguments{
\item{mri_path}{path to the volumetric T1 data.}

\item{out_dir}{optional output directory. Defaults to the same directory
as mri_path if not specified.}
}
\description{
Segment T1 weighted MRI data using the rpyANTs interface to ANTs and write to
file.
}
