% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{seq_cpmg_ideal}
\alias{seq_cpmg_ideal}
\title{CPMG style sequence with ideal pulses.}
\usage{
seq_cpmg_ideal(spin_params, ft, ref, TE = 0.03, echoes = 4)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{TE}{echo time in seconds.}

\item{echoes}{number of echoes.}
}
\value{
a list of resonance amplitudes and frequencies.
}
\description{
CPMG style sequence with ideal pulses.
}
