% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{seq_mega_press_ideal}
\alias{seq_mega_press_ideal}
\title{MEGA-PRESS sequence with ideal localisation pulses and Gaussian shaped
editing pulse.}
\usage{
seq_mega_press_ideal(
  spin_params,
  ft,
  ref,
  ed_freq = 1.89,
  TE1 = 0.015,
  TE2 = 0.053,
  BW = 110,
  steps = 50
)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{ed_freq}{editing pulse frequency in ppm.}

\item{TE1}{TE1 sequence parameter in seconds (TE=TE1+TE2).}

\item{TE2}{TE2 sequence parameter in seconds.}

\item{BW}{editing pulse bandwidth in Hz.}

\item{steps}{number of hard pulses used to approximate the editing pulse.}
}
\value{
list of resonance amplitudes and frequencies.
}
\description{
MEGA-PRESS sequence with ideal localisation pulses and Gaussian shaped
editing pulse.
}
