% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{seq_press_ideal}
\alias{seq_press_ideal}
\title{PRESS sequence with ideal pulses.}
\usage{
seq_press_ideal(spin_params, ft, ref, TE1 = 0.01, TE2 = 0.02)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{TE1}{TE1 sequence parameter in seconds.}

\item{TE2}{TE2 sequence parameter in seconds.}
}
\value{
a list of resonance amplitudes and frequencies.
}
\description{
PRESS sequence with ideal pulses.
}
