% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{seq_pulse_acquire}
\alias{seq_pulse_acquire}
\title{Simple pulse and acquire sequence with ideal pulses.}
\usage{
seq_pulse_acquire(spin_params, ft, ref, nuc = "1H", acq_delay = 0)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{nuc}{acquisition nucleus.}

\item{acq_delay}{delay between excitation and acquisition.}
}
\value{
list of resonance amplitudes and frequencies.
}
\description{
Simple pulse and acquire sequence with ideal pulses.
}
