% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{seq_slaser_ideal}
\alias{seq_slaser_ideal}
\title{sLASER sequence with ideal pulses.}
\usage{
seq_slaser_ideal(spin_params, ft, ref, TE1 = 0.008, TE2 = 0.009,
  TE3 = 0.011)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{TE1}{first echo time (between exc. and 1st echo) in seconds.}

\item{TE2}{second echo time (between 2nd echo and 4th echo) in seconds.}

\item{TE3}{third echo time (between 4th echo and 5th echo) in seconds.}
}
\value{
list of resonance amplitudes and frequencies.
}
\description{
sLASER sequence with ideal pulses.
}
