% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{set_ants_dir}
\alias{set_ants_dir}
\title{Set the ANTs installation directory location.}
\usage{
set_ants_dir(dir)
}
\arguments{
\item{dir}{ANTs installation directory.}
}
\description{
Set the ANTs installation directory location.
}
