% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{set_def_acq_paras}
\alias{set_def_acq_paras}
\title{Set the default acquisition parameters.}
\usage{
set_def_acq_paras(ft = getOption("spant.def_ft"),
  fs = getOption("spant.def_fs"), N = getOption("spant.def_N"),
  ref = getOption("spant.def_ref"))
}
\arguments{
\item{ft}{Trasmitter frequency in Hz.}

\item{fs}{Sampling frequency in Hz.}

\item{N}{Number of data points in the spectral dimension.}

\item{ref}{Reference value for ppm scale.}
}
\description{
Set the default acquisition parameters.
}
