% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{set_lcm_cmd}
\alias{set_lcm_cmd}
\title{Set the command to run the LCModel command-line program.}
\usage{
set_lcm_cmd(cmd)
}
\arguments{
\item{cmd}{oath to binary.}
}
\description{
Set the command to run the LCModel command-line program.
}
