% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{set_mask_xy_mat}
\alias{set_mask_xy_mat}
\title{Set the masked voxels in a 2D MRSI dataset to given spectrum.}
\usage{
set_mask_xy_mat(mrs_data, mask, mask_mrs_data)
}
\arguments{
\item{mrs_data}{MRSI data object.}

\item{mask}{matrix of boolean values specifying the voxels to set, where a
value of TRUE indicates the voxel should be set to mask_mrs_data.}

\item{mask_mrs_data}{the spectral data to be assigned to the masked voxels.}
}
\value{
updated dataset.
}
\description{
Set the masked voxels in a 2D MRSI dataset to given spectrum.
}
