% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{set_Ntrans}
\alias{set_Ntrans}
\title{Set the number of transients for an mrs_data object.}
\usage{
set_Ntrans(mrs_data, n_trans)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{n_trans}{number of acquired transients.}
}
\description{
Set the number of transients for an mrs_data object.
}
