% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp.R
\name{set_precomp_mode}
\alias{set_precomp_mode}
\title{Set the precompute mode.}
\usage{
set_precomp_mode(mode = NA)
}
\arguments{
\item{mode}{can be one of: "default", "overwrite", "clean" or "disabled".}
}
\description{
Set the precompute mode.
}
