% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precomp.R
\name{set_precomp_verbose}
\alias{set_precomp_verbose}
\title{Set the verbosity of the precompute function.}
\usage{
set_precomp_verbose(verbose = NA)
}
\arguments{
\item{verbose}{can be TRUE or FALSE.}
}
\description{
Set the verbosity of the precompute function.
}
