% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{set_ref}
\alias{set_ref}
\title{Set the ppm reference value (eg ppm value at 0Hz).}
\usage{
set_ref(mrs_data, ref)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{ref}{reference value for ppm scale.}
}
\description{
Set the ppm reference value (eg ppm value at 0Hz).
}
