% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{set_spec_values}
\alias{set_spec_values}
\title{Set spectral data points to a given value.}
\usage{
set_spec_values(mrs_data, xlim = NULL, value = 0, invert = FALSE)
}
\arguments{
\item{mrs_data}{data to be processed.}

\item{xlim}{spectral region (in ppm) of data points to set. If NULL (default)
all spectral points will be set.}

\item{value}{value to set the spectral region. Defaults to zero.}

\item{invert}{set all spectral points \emph{outside} the spectral region to the
given value.}
}
\value{
mrs_data object with modified spectral data points.
}
\description{
Set spectral data points to a given value.
}
