% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{set_tqn_cmd}
\alias{set_tqn_cmd}
\title{Set the command to run the TARQUIN command-line program.}
\usage{
set_tqn_cmd(cmd)
}
\arguments{
\item{cmd}{Path to binary.}
}
\description{
Set the command to run the TARQUIN command-line program.
}
