% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\name{set_tqn_path}
\alias{set_tqn_path}
\title{Set the path to the TARQUIN command-line program.}
\usage{
set_tqn_path(path)
}
\arguments{
\item{path}{Path to binary.}
}
\description{
Set the path to the TARQUIN command-line program.
}
