% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{set_tr}
\alias{set_tr}
\title{Set the repetition time of an MRS dataset.}
\usage{
set_tr(mrs_data, tr)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{tr}{repetition time in seonds.}
}
\value{
updated mrs_data set.
}
\description{
Set the repetition time of an MRS dataset.
}
