% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{shift_basis}
\alias{shift_basis}
\title{Apply frequency shifts to basis set signals.}
\usage{
shift_basis(basis, shifts)
}
\arguments{
\item{basis}{the basis to apply the shift to.}

\item{shifts}{a vector of frequency shifts to apply in ppm units. Must be the
same length as there are basis elements, or one value to be applied to all
elements.}
}
\value{
modified basis set object.
}
\description{
Apply frequency shifts to basis set signals.
}
