% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{sim_basis}
\alias{sim_basis}
\title{Simulate a basis set object.}
\usage{
sim_basis(mol_list, pul_seq = seq_pulse_acquire, ft = def_ft(),
  ref = def_ref(), fs = def_fs(), N = def_N(), xlim = NULL, ...)
}
\arguments{
\item{mol_list}{list of \code{mol_parameter} objects.}

\item{pul_seq}{pulse sequence function to use.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{fs}{sampling frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}

\item{xlim}{ppm range limiting signals to be simulated.}

\item{...}{extra parameters to pass to the pulse sequence function.}
}
\value{
basis object.
}
\description{
Simulate a basis set object.
}
