% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{sim_basis_mm_lip_lcm}
\alias{sim_basis_mm_lip_lcm}
\title{Simulate a macromolecular and lipid basis-set suitable for 1H brain MRS
analysis.}
\usage{
sim_basis_mm_lip_lcm(acq_paras = def_acq_paras())
}
\arguments{
\item{acq_paras}{list of acquisition parameters or an mrs_data object. See
\code{\link{def_acq_paras}}}
}
\value{
basis object.
}
\description{
Simulate a macromolecular and lipid basis-set suitable for 1H brain MRS
analysis.
}
