% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{sim_basis_tqn}
\alias{sim_basis_tqn}
\title{Simulate a basis file using TARQUIN.}
\usage{
sim_basis_tqn(fs = 2000, ft = 1.28e+08, N = 1024, ref = 4.65,
  opts = NULL)
}
\arguments{
\item{fs}{sampling frequency}

\item{ft}{transmitter frequency}

\item{N}{number of data points}

\item{ref}{chemical shift reference}

\item{opts}{List of options to pass to TARQUIN.}
}
\description{
Simulate a basis file using TARQUIN.
}
\examples{
\dontrun{
write_basis_tqn('test.basis',mrs_data,c("--echo","0.04"))
}
}
