% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{sim_mol}
\alias{sim_mol}
\title{Simulate a \code{mol_parameter} object.}
\usage{
sim_mol(
  mol,
  pul_seq = seq_pulse_acquire,
  ft = def_ft(),
  ref = def_ref(),
  fs = def_fs(),
  N = def_N(),
  xlim = NULL,
  ...
)
}
\arguments{
\item{mol}{\code{mol_parameter} object.}

\item{pul_seq}{pulse sequence function to use.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{fs}{sampling frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}

\item{xlim}{ppm range limiting signals to be simulated.}

\item{...}{extra parameters to pass to the pulse sequence function.}
}
\value{
\code{mrs_data} object.
}
\description{
Simulate a \code{mol_parameter} object.
}
