% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sim_noise}
\alias{sim_noise}
\title{Simulate a time-domain mrs_data object containing simulated Gaussian noise.}
\usage{
sim_noise(sd = 0.1, fs = def_fs(), ft = def_ft(), N = def_N(),
  ref = def_ref(), dyns = 1)
}
\arguments{
\item{sd}{standard deviation of the noise.}

\item{fs}{Sampling frequency in Hz.}

\item{ft}{Transmitter frequency in Hz.}

\item{N}{Number of data points in the spectral dimension.}

\item{ref}{Reference value for ppm scale.}

\item{dyns}{Number of dynamic scans to generate.}
}
\value{
mrs_data object.
}
\description{
Simulate a time-domain mrs_data object containing simulated Gaussian noise.
}
