% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_th_excit_profile}
\alias{sim_th_excit_profile}
\title{Simulate an ideal pulse excitation profile by smoothing a top-hat function
with a Gaussian.}
\usage{
sim_th_excit_profile(bw = 1500, sigma = 50, fa = 180)
}
\arguments{
\item{bw}{top-hat bandwidth (Hz).}

\item{sigma}{Gaussian width smoothing parameter (Hz).}

\item{fa}{intended flip angle of the pulse.}
}
\value{
data frame containing the frequency scale, excitation profile and
corresponding flip-angles.
}
\description{
Simulate an ideal pulse excitation profile by smoothing a top-hat function
with a Gaussian.
}
