% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sim_zero}
\alias{sim_zero}
\title{Simulate an mrs_data object containing complex zero valued samples.}
\usage{
sim_zero(fs = def_fs(), ft = def_ft(), N = def_N(), ref = def_ref(), dyns = 1)
}
\arguments{
\item{fs}{sampling frequency in Hz.}

\item{ft}{transmitter frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}

\item{ref}{reference value for ppm scale.}

\item{dyns}{number of dynamic scans to generate.}
}
\value{
mrs_data object.
}
\description{
Simulate an mrs_data object containing complex zero valued samples.
}
