% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{smooth_dyns}
\alias{smooth_dyns}
\title{Smooth data across the dynamic dimension with a Gaussian kernel.}
\usage{
smooth_dyns(mrs_data, sigma)
}
\arguments{
\item{mrs_data}{data to be smoothed.}

\item{sigma}{standard deviation of the underlying Gaussian kernel.}
}
\value{
smoothed mrs_data object.
}
\description{
Smooth data across the dynamic dimension with a Gaussian kernel.
}
