% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{sort_basis}
\alias{sort_basis}
\title{Sort the basis-set elements alphabetically.}
\usage{
sort_basis(basis)
}
\arguments{
\item{basis}{input basis.}
}
\value{
sorted basis.
}
\description{
Sort the basis-set elements alphabetically.
}
