% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\docType{package}
\name{spant-package}
\alias{spant}
\alias{spant-package}
\title{spant: spectroscopy analysis tools.}
\description{
spant provides a set of tools for reading, visualising and processing
Magnetic Resonance Spectroscopy (MRS) data.
}
\details{
To get started with spant, take a look at the introduction vignette:

\code{vignette("spant-intro", package="spant")}

Full list of vignettes:

\code{browseVignettes(package = "spant")}

Full list of functions:

\code{help(package = spant, help_type = "html")}

An online version of the documentation is available from:

\url{https://martin3141.github.io/spant/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://spantdoc.wilsonlab.co.uk/}
  \item \url{https://martin3141.github.io/spant/}
  \item \url{https://github.com/martin3141/spant}
  \item Report bugs at \url{https://github.com/martin3141/spant/issues}
}

}
\author{
\strong{Maintainer}: Martin Wilson \email{martin.wilson111@gmail.com} (\href{https://orcid.org/0000-0002-2089-3956}{ORCID})

Authors:
\itemize{
  \item Martin Wilson \email{martin.wilson111@gmail.com} (\href{https://orcid.org/0000-0002-2089-3956}{ORCID})
}

Other contributors:
\itemize{
  \item Yong Wang [contributor]
  \item John Muschelli [contributor]
}

}
