% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{spant_abfit_benchmark}
\alias{spant_abfit_benchmark}
\title{Simulate and fit some spectra with ABfit for benchmarking purposes. Basic
timing and performance metrics will be printed.}
\usage{
spant_abfit_benchmark(noise_reps = 10, return_res = FALSE, opts = abfit_opts())
}
\arguments{
\item{noise_reps}{number of spectra to fit with differing noise samples.}

\item{return_res}{return a list of fit_result objects.}

\item{opts}{ABfit options structure.}
}
\description{
Simulate and fit some spectra with ABfit for benchmarking purposes. Basic
timing and performance metrics will be printed.
}
