% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spant_mpress_drift}
\alias{spant_mpress_drift}
\title{Example MEGA-PRESS data with significant B0 drift.}
\format{
An object of class \code{mrs_data} of dimension 1 x 1 x 1 x 1 x 40 x 1 x 1024.
}
\usage{
spant_mpress_drift
}
\description{
Example MEGA-PRESS data with significant B0 drift.
}
\keyword{datasets}
