% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{spant_sim_fmrs_dataset}
\alias{spant_sim_fmrs_dataset}
\title{Simulate an example fMRS dataset for a block design fMRS experiment and
export a BIDS structure.}
\usage{
spant_sim_fmrs_dataset(output_dir = NULL)
}
\arguments{
\item{output_dir}{output directory for the BIDS data. Defaults to :
"HOME/sim_fmrs_dataset/data".}
}
\description{
Simulate an example fMRS dataset for a block design fMRS experiment and
export a BIDS structure.
}
