% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{spant_simulation_benchmark}
\alias{spant_simulation_benchmark}
\title{Simulate a typical metabolite basis set for benchmarking. Timing metrics will
be printed on completion.}
\usage{
spant_simulation_benchmark(sim_reps = 10, N = 1024)
}
\arguments{
\item{sim_reps}{number of times to simulate the basis set.}

\item{N}{number of FID data points to simulate.}
}
\description{
Simulate a typical metabolite basis set for benchmarking. Timing metrics will
be printed on completion.
}
