% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_sequences.R
\name{spin_echo_ideal}
\alias{spin_echo_ideal}
\title{Spin echo sequence with ideal pulses.}
\usage{
spin_echo_ideal(spin_params, ft, ref, TE = 0.03)
}
\arguments{
\item{spin_params}{spin system definition.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{TE}{echo time in seconds.}
}
\value{
a list of resonance amplitudes and frequencies.
}
\description{
Spin echo sequence with ideal pulses.
}
