% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{spin_sys}
\alias{spin_sys}
\title{Create a spin system object for pulse sequence simulation.}
\usage{
spin_sys(spin_params, ft, ref)
}
\arguments{
\item{spin_params}{an object describing the spin system properties.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}
}
\value{
spin system object.
}
\description{
Create a spin system object for pulse sequence simulation.
}
