% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{stackplot.fit_result}
\alias{stackplot.fit_result}
\title{Plot the fitting results of an object of class \code{fit_result} with
individual basis set components shown.}
\usage{
\method{stackplot}{fit_result}(x, xlim = NULL, y_offset = 1, dyn = 1,
  x_pos = 1, y_pos = 1, z_pos = 1, coil = 1, n = NULL,
  sub_bl = FALSE, labels = FALSE, ...)
}
\arguments{
\item{x}{fit_result object.}

\item{xlim}{the range of values to display on the x-axis, eg xlim = c(4,1).}

\item{y_offset}{separate basis signals in the y-axis direction by this value.}

\item{dyn}{the dynamic index to plot.}

\item{x_pos}{the x index to plot.}

\item{y_pos}{the y index to plot.}

\item{z_pos}{the z index to plot.}

\item{coil}{the coil element number to plot.}

\item{n}{single index element to plot (overrides other indices when given).}

\item{sub_bl}{subtract the baseline from the data and fit (logical).}

\item{labels}{print signal labels at the right side of the plot.}

\item{...}{further arguments to plot method.}
}
\description{
Plot the fitting results of an object of class \code{fit_result} with
individual basis set components shown.
}
