% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{stackplot}
\alias{stackplot}
\title{Produce a plot with multiple traces.}
\usage{
stackplot(x, ...)
}
\arguments{
\item{x}{object for plotting.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Produce a plot with multiple traces.
}
