% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sub_even_from_odd_dyns}
\alias{sub_even_from_odd_dyns}
\title{Subtract the even dynamic scans from the odd dynamic scans.}
\usage{
sub_even_from_odd_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}
}
\value{
subtracted result.
}
\description{
Subtract the even dynamic scans from the odd dynamic scans.
}
