% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sub_first_dyn}
\alias{sub_first_dyn}
\title{Subtract the first dynamic spectrum from a dynamic series.}
\usage{
sub_first_dyn(mrs_data, scale = 1)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{scale}{scale factor for the first spectrum.}
}
\value{
subtracted data.
}
\description{
Subtract the first dynamic spectrum from a dynamic series.
}
