% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sub_median_dyns}
\alias{sub_median_dyns}
\title{Subtract the median dynamic spectrum from a dynamic series.}
\usage{
sub_median_dyns(mrs_data, scale = 1)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{scale}{scale factor for the medium spectrum.}
}
\value{
subtracted data.
}
\description{
Subtract the median dynamic spectrum from a dynamic series.
}
