% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sub_odd_from_even_dyns}
\alias{sub_odd_from_even_dyns}
\title{Subtract the odd dynamic scans from the even dynamic scans.}
\usage{
sub_odd_from_even_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}
}
\value{
subtracted result.
}
\description{
Subtract the odd dynamic scans from the even dynamic scans.
}
