% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sum_dyns}
\alias{sum_dyns}
\title{Calculate the sum of data dynamics.}
\usage{
sum_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}
}
\value{
Sum of data dynamics.
}
\description{
Calculate the sum of data dynamics.
}
