% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sum_mrs}
\alias{sum_mrs}
\title{Sum two mrs_data objects.}
\usage{
sum_mrs(a, b, force = FALSE)
}
\arguments{
\item{a}{first mrs_data object to be summed.}

\item{b}{second mrs_data object to be summed.}

\item{force}{set to TRUE to force mrs_data objects to be summed, even if they
are in different time/frequency domains.}
}
\value{
a + b
}
\description{
Sum two mrs_data objects.
}
