% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sum_mrs_list}
\alias{sum_mrs_list}
\title{Return the sum of a list of mrs_data objects.}
\usage{
sum_mrs_list(mrs_list)
}
\arguments{
\item{mrs_list}{list of mrs_data objects.}
}
\value{
sum \code{mrs_data} object.
}
\description{
Return the sum of a list of mrs_data objects.
}
