% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sv_res_table}
\alias{sv_res_table}
\title{Output a table of fit amplitudes and error estimates for a single-voxel fit.}
\usage{
sv_res_table(fit_res, format_out = FALSE)
}
\arguments{
\item{fit_res}{input vector.}

\item{format_out}{reduce the accuracy of values to aid table formatting.}
}
\value{
data.frame of values.
}
\description{
Output a table of fit amplitudes and error estimates for a single-voxel fit.
}
