% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs_batch_fit.R
\name{svs_1h_brain_analysis_new}
\alias{svs_1h_brain_analysis_new}
\title{Standard SVS 1H brain analysis pipeline.}
\usage{
svs_1h_brain_analysis_new(
  metab,
  w_ref = NULL,
  p_vols = NULL,
  dfp_corr = TRUE,
  omit_bad_dynamics = TRUE,
  basis = NULL,
  te = NULL,
  tr = NULL
)
}
\arguments{
\item{metab}{filepath or mrs_data object containing MRS metabolite data.}

\item{w_ref}{filepath or mrs_data object containing MRS water reference data.}

\item{p_vols}{a numeric vector of partial volumes expressed as percentages.
Defaults to 100\% white matter. A voxel containing 100\% gray matter tissue
would use : p_vols = c(WM = 0, GM = 100, CSF = 0).}

\item{dfp_corr}{perform dynamic frequency and phase correction using the RATS
method.}

\item{omit_bad_dynamics}{detect and remove bad dynamics.}

\item{basis}{basis set object to use for analysis.}

\item{te}{metabolite mrs data echo time in seconds. If not supplied this will
be guessed from the metab data file.}

\item{tr}{metabolite mrs data repetition time in seconds. If not supplied
this will be guessed from the metab data file.}
}
\description{
Standard SVS 1H brain analysis pipeline.
}
