% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{td2fd}
\alias{td2fd}
\title{Transform time-domain data to the frequency-domain.}
\usage{
td2fd(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data in time-domain representation.}
}
\value{
MRS data in frequency-domain representation.
}
\description{
Transform time-domain data to the frequency-domain.
}
