% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{td_conv_filt}
\alias{td_conv_filt}
\title{Time-domain convolution based filter.}
\usage{
td_conv_filt(mrs_data, K = 25, ext = 1)
}
\arguments{
\item{mrs_data}{MRS data to be filtered.}

\item{K}{window width in data points.}

\item{ext}{point seperation for linear extrapolation.}
}
\description{
Time-domain convolution based filter described by:
Marion D, Ikura M, Bax A. Improved solvent suppression in one-dimensional and
twodimensional NMR spectra by convolution of time-domain data. J Magn Reson
1989;84:425-430.
}
