% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{te}
\alias{te}
\title{Return the echo time of an MRS dataset.}
\usage{
te(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
echo time in seconds.
}
\description{
Return the echo time of an MRS dataset.
}
