% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{tr}
\alias{tr}
\title{Return the repetition time of an MRS dataset.}
\usage{
tr(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
repetition time in seconds.
}
\description{
Return the repetition time of an MRS dataset.
}
