% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_paths}
\alias{trim_paths}
\title{Trim a vector of filesystem paths.}
\usage{
trim_paths(paths, dir = 0, char = 0)
}
\arguments{
\item{paths}{vectors of filesystem paths.}

\item{dir}{number of times to apply the base dirname function.}

\item{char}{number of characters to trim from the end of each path. Note this
is performed after the dir based trimming.}
}
\value{
a vector of trimmed paths.
}
\description{
Trim a vector of filesystem paths.
}
