% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{vec2mrs_data}
\alias{vec2mrs_data}
\title{Convert a vector into a mrs_data object.}
\usage{
vec2mrs_data(
  vec,
  fs = def_fs(),
  ft = def_ft(),
  ref = def_ref(),
  dyns = 1,
  fd = FALSE
)
}
\arguments{
\item{vec}{the data vector.}

\item{fs}{sampling frequency in Hz.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{dyns}{replicate the data across the dynamic dimension.}

\item{fd}{flag to indicate if the matrix is in the frequency domain (logical).}
}
\value{
mrs_data object.
}
\description{
Convert a vector into a mrs_data object.
}
