% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{write_basis_niidir}
\alias{write_basis_niidir}
\title{Write a basis object to folder containing one NIfTI MRS file for each basis
element.}
\usage{
write_basis_niidir(basis, basis_dir)
}
\arguments{
\item{basis}{basis object to be exported.}

\item{basis_dir}{directory for basis files to be written.}
}
\description{
Write a basis object to folder containing one NIfTI MRS file for each basis
element.
}
