% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{write_basis_tqn}
\alias{write_basis_tqn}
\title{Generate a basis file using TARQUIN.}
\usage{
write_basis_tqn(basis_file, metab_data, opts = NULL)
}
\arguments{
\item{basis_file}{The filename of the basis file to be generated.}

\item{metab_data}{The MRS data object to match the generated basis parameters.}

\item{opts}{List of options to pass to TARQUIN.}
}
\description{
Generate a basis file using TARQUIN.
}
\examples{
\dontrun{
write_basis_tqn('test.basis',mrs_data,c("--echo","0.04"))
}
}
