% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_io.R
\name{write_mrs_dpt_v2}
\alias{write_mrs_dpt_v2}
\title{Write MRS data object to file in dangerplot (dpt) v2 format.}
\usage{
write_mrs_dpt_v2(fname, mrs_data)
}
\arguments{
\item{fname}{the filename of the output dpt format MRS data.}

\item{mrs_data}{object to be written to file.}
}
\description{
Write MRS data object to file in dangerplot (dpt) v2 format.
}
\examples{
\dontrun{
mrs_data <- write_mrs_dpt_v2("my_mrs_data.dpt", my_mrs_data)
}
}
