% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_io.R
\name{write_mrs_lcm_raw}
\alias{write_mrs_lcm_raw}
\title{Write MRS data object to file in a RAW format compatible with LCModel.}
\usage{
write_mrs_lcm_raw(fname, mrs_data)
}
\arguments{
\item{fname}{the filename of the output RAW format MRS data.}

\item{mrs_data}{object to be written to file.}
}
\description{
Write MRS data object to file in a RAW format compatible with LCModel.
}
\examples{
\dontrun{
mrs_data <- write_mrs_lcm_raw("my_mrs_data.RAW", my_mrs_data)
}
}
