% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_write_nifti.R
\name{write_mrs_nifti}
\alias{write_mrs_nifti}
\title{Write MRS data object to file in NIFTI format.}
\usage{
write_mrs_nifti(mrs_data, fname)
}
\arguments{
\item{mrs_data}{object to be written to file.}

\item{fname}{the filename of the output NIFTI MRS data.}
}
\description{
Write MRS data object to file in NIFTI format.
}
