% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_io.R
\name{write_mrs_rds}
\alias{write_mrs_rds}
\title{Write MRS data object to file in rds format.}
\usage{
write_mrs_rds(fname, mrs_data)
}
\arguments{
\item{fname}{the filename of the output rds MRS data.}

\item{mrs_data}{object to be written to file.}
}
\description{
Write MRS data object to file in rds format.
}
