% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_shapes.R
\name{write_pulse_ascii}
\alias{write_pulse_ascii}
\title{Write an ASCII formatted pulse file.}
\usage{
write_pulse_ascii(pulse, path)
}
\arguments{
\item{pulse}{pulse data object.}

\item{path}{file path for export.}
}
\description{
Write an ASCII formatted pulse file.
}
