% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zero_fade_spec}
\alias{zero_fade_spec}
\title{Fade a spectrum to zero by frequency domain multiplication with a tanh
function. Note this operation distorts data points at the end of the FID.}
\usage{
zero_fade_spec(mrs_data, start_ppm, end_ppm)
}
\arguments{
\item{mrs_data}{data to be faded.}

\item{start_ppm}{start point of the fade in ppm units.}

\item{end_ppm}{end point of the fade in ppm units.}
}
\value{
modified mrs_data object.
}
\description{
Fade a spectrum to zero by frequency domain multiplication with a tanh
function. Note this operation distorts data points at the end of the FID.
}
