% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{zero_nzoc}
\alias{zero_nzoc}
\title{Zero all non-zero-order coherences}
\usage{
zero_nzoc(sys, rho)
}
\arguments{
\item{sys}{spin system object}

\item{rho}{density matrix}
}
\value{
density matrix
}
\description{
Zero all non-zero-order coherences
}
