% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{zero_pqcs}
\alias{zero_pqcs}
\title{Zero the positive quantum coherences of a density matrix}
\usage{
zero_pqcs(sys, rho)
}
\arguments{
\item{sys}{spin system object}

\item{rho}{density matrix}
}
\value{
density matrix
}
\description{
Zero the positive quantum coherences of a density matrix
}
