% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zero_td_pts_end}
\alias{zero_td_pts_end}
\title{Set \code{mrs_data} object data points at the end of the FID to zero.}
\usage{
zero_td_pts_end(mrs_data, pts)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{pts}{number of end points to set to zero.}
}
\value{
modified \code{mrs_data} object.
}
\description{
Set \code{mrs_data} object data points at the end of the FID to zero.
}
