% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zf_xy}
\alias{zf_xy}
\title{Zero-fill MRSI data in the k-space x-y direction.}
\usage{
zf_xy(mrs_data, factor = 2)
}
\arguments{
\item{mrs_data}{MRSI data.}

\item{factor}{zero-filling factor, a factor of 2 returns a dataset with
twice the original points in the x-y directions. Factors smaller than one
are permitted, such that a factor of 0.5 returns half the k-space points in
the x-y directions.}
}
\value{
zero-filled data.
}
\description{
Zero-fill MRSI data in the k-space x-y direction.
}
