% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abfit.R
\name{abfit_opts}
\alias{abfit_opts}
\title{Return a list of options for an ABFit analysis.}
\usage{
abfit_opts(
  init_damping = 5,
  maxiters = 1024,
  max_shift = 10,
  max_damping = 15,
  max_phase = 360,
  lambda = NULL,
  ppm_left = 4,
  ppm_right = 0.2,
  zp = TRUE,
  bl_ed_pppm = 2,
  auto_bl_flex = TRUE,
  bl_comps_pppm = 15,
  export_sp_fit = FALSE,
  max_asym = 0.25,
  max_basis_shift = 1,
  max_basis_damping = 2,
  maxiters_pre = 1000,
  algo_pre = "NLOPT_LN_NELDERMEAD",
  min_bl_ed_pppm = NULL,
  max_bl_ed_pppm = 7,
  auto_bl_flex_n = 20,
  pre_fit_bl_ed_pppm = 1,
  remove_lip_mm_prefit = FALSE,
  pre_align = TRUE,
  max_pre_align_shift = 0.1,
  pre_align_ref_freqs = c(2.01, 3.03, 3.22),
  noise_region = c(-0.5, -2.5),
  optimal_smooth_criterion = "maic",
  aic_smoothing_factor = 5,
  anal_jac = TRUE,
  pre_fit_ppm_left = 4,
  pre_fit_ppm_right = 1.8,
  phi1_optim = FALSE,
  phi1_init = 0,
  max_dphi1 = 0.2,
  max_basis_shift_broad = 1,
  max_basis_damping_broad = 2
)
}
\arguments{
\item{init_damping}{initial value of the Gaussian global damping parameter
(Hz). Very poorly shimmed or high field data may benefit from a larger value.}

\item{maxiters}{The maximum number of iterations to run for the detailed fit.}

\item{max_shift}{The maximum allowable shift to be applied in the
opimisation phase of fitting (Hz).}

\item{max_damping}{maximum permitted value of the global damping parameter
(Hz).}

\item{max_phase}{maximum permitted value of the global zero-order phase term
(degrees).}

\item{lambda}{todo.}

\item{ppm_left}{downfield frequency limit for the fitting range (ppm).}

\item{ppm_right}{upfield frequency limit for the fitting range (ppm).}

\item{zp}{zero pad the data to twice the original length before fitting.}

\item{bl_ed_pppm}{todo.}

\item{auto_bl_flex}{todo.}

\item{bl_comps_pppm}{todo.}

\item{export_sp_fit}{add the fitted spline functions to the fit result.}

\item{max_asym}{maximum allowable value of the asymetry parameter.}

\item{max_basis_shift}{maximum allowable frequency shift for individual basis
signals (Hz).}

\item{max_basis_damping}{maximum allowable Lorentzian damping factor for
individual basis signals (Hz).}

\item{maxiters_pre}{maximum iterations for the coarse (pre-)fit.}

\item{algo_pre}{optimisation method for the coarse (pre-)fit.}

\item{min_bl_ed_pppm}{todo.}

\item{max_bl_ed_pppm}{todo.}

\item{auto_bl_flex_n}{todo.}

\item{pre_fit_bl_ed_pppm}{todo.}

\item{remove_lip_mm_prefit}{todo.}

\item{pre_align}{todo.}

\item{max_pre_align_shift}{todo.}

\item{pre_align_ref_freqs}{todo.}

\item{noise_region}{todo.}

\item{optimal_smooth_criterion}{todo.}

\item{aic_smoothing_factor}{todo.}

\item{anal_jac}{todo.}

\item{pre_fit_ppm_left}{todo.}

\item{pre_fit_ppm_right}{todo.}

\item{phi1_optim}{todo.}

\item{phi1_init}{todo.}

\item{max_dphi1}{todo.}

\item{max_basis_shift_broad}{todo.}

\item{max_basis_damping_broad}{todo.}
}
\value{
full list of options.
}
\description{
Return a list of options for an ABFit analysis.
}
\examples{
opts <- abfit_opts(ppm_left = 4.2)
}
