% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{apodise_xy}
\alias{apodise_xy}
\title{Apodise MRSI data in the x-y direction with a k-space hamming filter.}
\usage{
apodise_xy(mrs_data)
}
\arguments{
\item{mrs_data}{MRSI data.}
}
\value{
apodised data.
}
\description{
Apodise MRSI data in the x-y direction with a k-space hamming filter.
}
