% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{apply_pvc}
\alias{apply_pvc}
\title{Apply partial volume correction to a fitting result object.}
\usage{
apply_pvc(fit_result, p_vols, tr, te)
}
\arguments{
\item{fit_result}{a \code{fit_result} object to apply partial volume
correction.}

\item{p_vols}{a numeric vector of partial volumes.}

\item{tr}{the MRS TR.}

\item{te}{the MRS TE.}
}
\value{
a \code{fit_result} object with an added results table:
"results_pvc".
}
\description{
Apply partial volume correction to a fitting result object.
}
