% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{back_extrap}
\alias{back_extrap}
\title{Back extrapolate time-domain points using the Burg autoregressive model}
\usage{
back_extrap(mrs_data, extrap_pts, pred_pts = NA)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{extrap_pts}{number of points to extrapolate.}

\item{pred_pts}{number of points to base the extrapolation on.}
}
\value{
back extrapolated data.
}
\description{
Back extrapolate time-domain points using the Burg autoregressive model
}
