% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{bc_als}
\alias{bc_als}
\title{Baseline correction using the ALS method.}
\usage{
bc_als(mrs_data, lambda = 10000, p = 0.001)
}
\arguments{
\item{mrs_data}{mrs_data object.}

\item{lambda}{lambda parameter.}

\item{p}{p parameter.}
}
\value{
baseline corrected data.
}
\description{
Baseline correction using the ALS method.
}
