% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{circ_mask}
\alias{circ_mask}
\title{Create a logical circular mask spanning the full extent of an n x n matrix.}
\usage{
circ_mask(d, n, offset = 1)
}
\arguments{
\item{d}{diameter of the mask.}

\item{n}{number of matrix rows and columns.}

\item{offset}{offset the mask centre in matrix dimension units.}
}
\value{
logical n x n mask matrix.
}
\description{
Create a logical circular mask spanning the full extent of an n x n matrix.
}
