% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{collapse_to_dyns}
\alias{collapse_to_dyns}
\title{Collapse MRS data by concatenating spectra along the dynamic dimension.}
\usage{
collapse_to_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{mrs_data object to be collapsed.}
}
\value{
collapsed data with spectra concatentated along the dynamic
dimension.
}
\description{
Collapse MRS data by concatenating spectra along the dynamic dimension.
}
