% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{comb_coils}
\alias{comb_coils}
\title{Combine coil data based on the first data point of a reference signal.}
\usage{
comb_coils(metab_mrs, ref_mrs = NULL, noise_mrs = NULL, scale = TRUE,
  sum_coils = TRUE)
}
\arguments{
\item{metab_mrs}{MRS data containing metabolite data.}

\item{ref_mrs}{MRS data containing reference data (optional).}

\item{noise_mrs}{MRS data from a noise scan (optional).}

\item{scale}{option to rescale coil elements based on the first data point
(logical).}

\item{sum_coils}{sum the coil elements as a final step (logical).}
}
\value{
MRS data.
}
\description{
By default, elements are phased and scaled prior to summation. Where a
reference signal is not given, the mean dynamic signal will be used
instead.
}
