% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{diff.mrs_data}
\alias{diff.mrs_data}
\title{Apply the diff operator to an MRS dataset.}
\usage{
\method{diff}{mrs_data}(x, lag = 1, order = 1)
}
\arguments{
\item{x}{MRS data.}

\item{lag}{diff lag (default = 1).}

\item{order}{diff order (default = 1).}
}
\value{
MRS data following diff operator.
}
\description{
Apply the diff operator to an MRS dataset.
}
