% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{dyn_acq_times}
\alias{dyn_acq_times}
\title{Return a time scale vector of acquisition times for a dynamic MRS scan. The
first temporal scan is assigned a value of 0.}
\usage{
dyn_acq_times(mrs_data)
}
\arguments{
\item{mrs_data}{MRS data.}
}
\value{
time scale vector in units of seconds.
}
\description{
Return a time scale vector of acquisition times for a dynamic MRS scan. The
first temporal scan is assigned a value of 0.
}
