% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{ecc}
\alias{ecc}
\title{Eddy current correction.}
\usage{
ecc(metab, ref, rev = FALSE)
}
\arguments{
\item{metab}{MRS data to be corrected.}

\item{ref}{ceference dataset.}

\item{rev}{reverse the correction.}
}
\value{
corrected data in the time domain.
}
\description{
Apply eddy current correction using the Klose method.
}
\details{
In vivo proton spectroscopy in presence of eddy currents.
Klose U.
Magn Reson Med. 1990 Apr;14(1):26-30.
}
