% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svs.R
\name{fit_svs}
\alias{fit_svs}
\title{Standard SVS 1H brain analysis pipeline.}
\usage{
fit_svs(
  metab,
  w_ref = NULL,
  output_dir = NULL,
  basis = NULL,
  p_vols = NULL,
  format = NULL,
  pul_seq = NULL,
  TE = NULL,
  TR = NULL,
  TE1 = NULL,
  TE2 = NULL,
  TE3 = NULL,
  TM = NULL,
  append_basis = NULL,
  remove_basis = NULL,
  dfp_corr = TRUE,
  output_ratio = "tCr",
  ecc = FALSE,
  fit_opts = NULL,
  legacy_ws = FALSE,
  w_att = 0.7,
  w_conc = 35880,
  verbose = FALSE
)
}
\arguments{
\item{metab}{path or mrs_data object containing MRS metabolite data.}

\item{w_ref}{path or mrs_data object containing MRS water reference data.}

\item{output_dir}{directory path to output fitting results.}

\item{basis}{precompiled basis set object to use for analysis.}

\item{p_vols}{a numeric vector of partial volumes expressed as percentages.
Defaults to 100\% white matter. A voxel containing 100\% gray matter tissue
would use : p_vols = c(WM = 0, GM = 100, CSF = 0).}

\item{format}{Override automatic data format detection. See format argument
in \code{\link[=read_mrs]{read_mrs()}} for permitted values.}

\item{pul_seq}{Pulse sequence to use for basis simulation. Can be one of the
following values : "press", "press_ideal", "press_shaped", "steam" or
"slaser". If "press" then "press_ideal" will be assumed unless the magnetic
field is stronger that 2.8 Tesla, "press_shaped" will be assumed for 2.9
Tesla and above.}

\item{TE}{metabolite mrs data echo time in seconds. If not supplied this will
be guessed from the metab data file.}

\item{TR}{metabolite mrs data repetition time in seconds. If not supplied
this will be guessed from the metab data file.}

\item{TE1}{PRESS or sLASER sequence timing parameter in seconds.}

\item{TE2}{PRESS or sLASER sequence timing parameter in seconds.}

\item{TE3}{sLASER sequence timing parameter in seconds.}

\item{TM}{STEAM mixing time parameter in seconds.}

\item{append_basis}{names of extra signals to add to the default basis. Eg
append_basis = c("peth", "cit"). Cannot be used with precompiled basis sets.}

\item{remove_basis}{grep expression to match names of signals to remove from
the basis. For example: use "*" to remove all signals, "^mm|^lip" to remove
all macromolecular and lipid signals, "^lac" to remove lactate. This operation
is performed before signals are added with append_basis. Cannot be used with
precompiled basis sets.}

\item{dfp_corr}{perform dynamic frequency and phase correction using the RATS
method.}

\item{output_ratio}{optional string to specify a metabolite ratio to output.
Defaults to "tCr" and multiple metabolites may be specified for multiple
outputs. Set as NULL to omit.}

\item{ecc}{option to perform water reference based eddy current correction,
defaults to FALSE.}

\item{fit_opts}{options to pass to ABfit.}

\item{legacy_ws}{perform and output legacy water scaling compatible with
default LCModel and TARQUIN behaviour. See w_att and w_conc arguments to
change the default assumptions. Default value is FALSE.}

\item{w_att}{water attenuation factor (default = 0.7) for legacy water
scaling. Assumes water T2 of 80ms and a TE = 30 ms. exp(-30ms / 80ms) ~ 0.7.}

\item{w_conc}{assumed water concentration (default = 35880) for legacy water
scaling. Default value corresponds to typical white matter. Set to 43300 for
gray matter, and 55556 for phantom measurements.}

\item{verbose}{output potentially useful information.}
}
\description{
Note this function is still under development and liable to changes.
}
\examples{
metab <- system.file("extdata", "philips_spar_sdat_WS.SDAT",
                     package = "spant")
w_ref <- system.file("extdata", "philips_spar_sdat_W.SDAT",
                     package = "spant")
\dontrun{
fit_result <- svs_1h_brain_analysis(metab, w_ref, "fit_res_dir")
}
}
