% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svs.R
\name{fit_svs_group_results}
\alias{fit_svs_group_results}
\title{Combine fitting results for group analysis.}
\usage{
fit_svs_group_results(search_path, output_dir = "fit_svs_group_results")
}
\arguments{
\item{search_path}{path to start recursive search for fitting results.}

\item{output_dir}{directory path to store group results.}
}
\description{
Combine fitting results for group analysis.
}
