% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_svs.R
\name{fit_svs_group_results}
\alias{fit_svs_group_results}
\title{Combine fitting results for group analysis.}
\usage{
fit_svs_group_results(
  search_path = NULL,
  paths = NULL,
  output_dir = "fit_svs_group_results",
  verbose = TRUE
)
}
\arguments{
\item{search_path}{path to start recursive search for fitting results.
Cannot be used together with the paths argument.}

\item{paths}{a set of paths to spant output files, usually named :
"spant_fit_svs_data.rds". Cannot be used together with the search_path
argument.}

\item{output_dir}{directory path to store group results.}

\item{verbose}{verbose, defaults to TRUE.}
}
\description{
Combine fitting results for group analysis.
}
