% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{gen_numeric_reg}
\alias{gen_numeric_reg}
\title{Generate a regressor from a numeric vector.}
\usage{
gen_numeric_reg(
  in_vec,
  name,
  mrs_data = NULL,
  tr = NULL,
  Ndyns = NULL,
  Ntrans = NULL
)
}
\arguments{
\item{in_vec}{a numeric input vector the same length as the number of stored
dynamic scans.}

\item{name}{a character vector representing the regressor name.}

\item{mrs_data}{mrs_data object for timing information.}

\item{tr}{repetition time.}

\item{Ndyns}{number of dynamic scans stored, potentially less than Ntrans
if block averaging has been performed.}

\item{Ntrans}{number of dynamic scans acquired.}
}
\value{
a single baseline regressor with value of 1.
}
\description{
Generate a regressor from a numeric vector.
}
