% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{get_1h_brain_basis_names}
\alias{get_1h_brain_basis_names}
\title{Return a character vector of common 1H molecules found in healthy human
brain.}
\usage{
get_1h_brain_basis_names(extra = NULL, remove = NULL, inc_lip_mm = TRUE)
}
\arguments{
\item{extra}{optional character vector of additional molecular names. Eg
c("asc", "gly", "peth").}

\item{remove}{optional character vector of molecular names to remove from the
set. Eg c("m_cr_ch2").}

\item{inc_lip_mm}{include Lipid and MM basis signals.}
}
\value{
a character vector of molecule names.
}
\description{
Note, this is a basic set and it may be appropriate to also include Asc,
Gly and PEth for high quality MRS data.
}
