% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_seg_ind}
\alias{get_seg_ind}
\title{Get the indices of data points lying between two values.}
\usage{
get_seg_ind(scale, start, end)
}
\arguments{
\item{scale}{the full list of values.}

\item{start}{the smallest value in the subset.}

\item{end}{the largest value in the subset.}
}
\value{
a set of indices.
}
\description{
Get the indices of data points lying between two values.
}
