% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{get_subset}
\alias{get_subset}
\title{Extract a subset of MRS data.}
\usage{
get_subset(
  mrs_data,
  x_set = NULL,
  y_set = NULL,
  z_set = NULL,
  dyn_set = NULL,
  coil_set = NULL,
  fd_set = NULL,
  td_set = NULL
)
}
\arguments{
\item{mrs_data}{MRS data object.}

\item{x_set}{x indices to include in the output (default all).}

\item{y_set}{y indices to include in the output (default all).}

\item{z_set}{z indices to include in the output (default all).}

\item{dyn_set}{dynamic indices to include in the output (default all).}

\item{coil_set}{coil indices to include in the output (default all).}

\item{fd_set}{frequency domain data indices to include in the output (default
all).}

\item{td_set}{time-domain indices to include in the output (default all).}
}
\value{
selected subset of MRS data.
}
\description{
Extract a subset of MRS data.
}
