% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{image.mrs_data}
\alias{image.mrs_data}
\title{TODO document.}
\usage{
\method{image}{mrs_data}(x, mode = "real", xlim = NULL, col = NULL,
  dim = "dyn", x_pos = NA, y_pos = NA, z_pos = NA, dyn = 1,
  coil = 1, ...)
}
\description{
TODO document.
}
