% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{image.mrs_data}
\alias{image.mrs_data}
\title{Image plot method for objects of class mrs_data.}
\usage{
\method{image}{mrs_data}(x, xlim = NULL, mode = "real", col = NULL,
  dim = "dyn", x_pos = NULL, y_pos = NULL, z_pos = NULL, dyn = 1,
  coil = 1, ...)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{xlim}{the range of values to display on the x-axis, eg xlim = c(4,1).}

\item{mode}{representation of the complex numbers to be plotted, can be one
of: "real", "imag" or "abs".}

\item{col}{Colour map to use, defaults to viridis if the package is
available.}

\item{dim}{the dimension to display on the y-axis, can be one of: "dyn", "x",
"y", "z" or "coil".}

\item{x_pos}{the x index to plot.}

\item{y_pos}{the y index to plot.}

\item{z_pos}{the z index to plot.}

\item{dyn}{the dynamic index to plot.}

\item{coil}{the coil element number to plot.}

\item{...}{other arguments to pass to the plot method.}
}
\description{
Image plot method for objects of class mrs_data.
}
