% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ants_utils.R
\name{install_ants}
\alias{install_ants}
\title{Install ANTs / ANTsX from : https://github.com/ANTsX/ANTs/releases}
\usage{
install_ants(platform, version = "2.6.5")
}
\arguments{
\item{platform}{see the releases page for supported platforms. Common
platforms include : "ubuntu-24.04-X64-gcc" or "macos-14-ARM64-clang".}

\item{version}{ANTs version to install, defaults to "2.6.5".}
}
\description{
Install ANTs / ANTsX from : https://github.com/ANTsX/ANTs/releases
}
