% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{int_spec}
\alias{int_spec}
\title{Integrate a spectral region.}
\usage{
int_spec(mrs_data, xlim = NULL, scale = "ppm", mode = "real")
}
\arguments{
\item{mrs_data}{MRS data.}

\item{xlim}{Spectral range to be integrated.}

\item{scale}{Units of xlim, can be : "ppm", "Hz" or "points".}

\item{mode}{Spectral mode, can be : "real", "imag" or "abs".}
}
\value{
An array of integral values.
}
\description{
Integrate a spectral region.
}
