% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{int_spec}
\alias{int_spec}
\title{Integrate a spectral region.}
\usage{
int_spec(mrs_data, xlim = NULL, scale = "ppm", mode = "re", summation = "sum")
}
\arguments{
\item{mrs_data}{MRS data.}

\item{xlim}{spectral range to be integrated (defaults to full range).}

\item{scale}{units of xlim, can be : "ppm", "hz" or "points".}

\item{mode}{spectral mode, can be : "re", "im", "mod" or "cplx".}

\item{summation}{can be "sum" (default), "mean" or "l2".}
}
\value{
an array of integral values.
}
\description{
Integrate a spectral region.
}
