% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{inv_even_dyns}
\alias{inv_even_dyns}
\title{Invert even numbered dynamic scans starting from 1 (2,4,6...).}
\usage{
inv_even_dyns(mrs_data)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}
}
\value{
dynamic MRS data with inverted even numbered scans.
}
\description{
Invert even numbered dynamic scans starting from 1 (2,4,6...).
}
