% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{l2_reg}
\alias{l2_reg}
\title{Perform l2 regularisation artefact suppression using the method proposed by
Bilgic et al. JMRI 40(1):181-91 2014.}
\usage{
l2_reg(mrs_data, A, b)
}
\arguments{
\item{mrs_data}{input data for artefact suppression.}

\item{A}{matrix of spectral data points containing the artefact basis
signals.}

\item{b}{regularisation parameter.}
}
\value{
l2 reconstructed mrs_data object.
}
\description{
Perform l2 regularisation artefact suppression using the method proposed by
Bilgic et al. JMRI 40(1):181-91 2014.
}
