% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mask_xy}
\alias{mask_xy}
\title{Mask an MRSI dataset in the x-y direction}
\usage{
mask_xy(mrs_data, x_dim, y_dim)
}
\arguments{
\item{mrs_data}{MRS data object.}

\item{x_dim}{x dimension output length.}

\item{y_dim}{y dimension output length.}
}
\value{
masked MRS data.
}
\description{
Mask an MRSI dataset in the x-y direction
}
