% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mat2mrs_data}
\alias{mat2mrs_data}
\title{Convert a matrix (with spectral points in the column dimension and dynamics
in the row dimensions) into a mrs_data object.}
\usage{
mat2mrs_data(mat, fs = def_fs(), ft = def_ft(), ref = def_ref(), fd = FALSE)
}
\arguments{
\item{mat}{data matrix.}

\item{fs}{sampling frequency in Hz.}

\item{ft}{transmitter frequency in Hz.}

\item{ref}{reference value for ppm scale.}

\item{fd}{flag to indicate if the matrix is in the frequency domain (logical).}
}
\value{
mrs_data object.
}
\description{
Convert a matrix (with spectral points in the column dimension and dynamics
in the row dimensions) into a mrs_data object.
}
