% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{mean_dyn_blocks}
\alias{mean_dyn_blocks}
\title{Calculate the mean of adjacent dynamic scans.}
\usage{
mean_dyn_blocks(mrs_data, block_size)
}
\arguments{
\item{mrs_data}{Dynamic MRS data.}

\item{block_size}{Number of adjacent dynamics scans to average over.}
}
\value{
Dynamic data averaged in blocks.
}
\description{
Calculate the mean of adjacent dynamic scans.
}
