% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{norm_mrs}
\alias{norm_mrs}
\title{Normalise mrs_data to a spectral region.}
\usage{
norm_mrs(mrs_data, xlim = NULL, scale = "ppm", mode = "re",
  summation = "l2")
}
\arguments{
\item{mrs_data}{MRS data.}

\item{xlim}{spectral range to be integrated (defaults to full range).}

\item{scale}{units of xlim, can be : "ppm", "Hz" or "points".}

\item{mode}{spectral mode, can be : "re", "im" or "mod".}

\item{summation}{can be "simple" or "l2" (default).}
}
\value{
normalised data.
}
\description{
Normalise mrs_data to a spectral region.
}
