% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{one_page_pdf}
\alias{one_page_pdf}
\title{Export a one-page pdf of a single fit result}
\usage{
one_page_pdf(fit_res, pdf_out_path, title = NULL)
}
\arguments{
\item{fit_res}{\code{fit_result} object.}

\item{pdf_out_path}{path to the exported pdf file.}

\item{title}{ouptut title.}
}
\description{
Export a one-page pdf of a single fit result
}
