% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\name{phase_ref_1h_brain}
\alias{phase_ref_1h_brain}
\title{Corrected zero order phase and chemical shift offset in 1H MRS data from the
brain.}
\usage{
phase_ref_1h_brain(mrs_data, mean_ref = FALSE)
}
\arguments{
\item{mrs_data}{MRS data to be corrected.}

\item{mean_ref}{apply the phase and offset of the mean spectrum to all
others. Default is FALSE.}
}
\value{
corrected MRS data.
}
\description{
Corrected zero order phase and chemical shift offset in 1H MRS data from the
brain.
}
