% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{plot.fit_result}
\alias{plot.fit_result}
\title{Plot the fitting results of an object of class \code{fit_result}.}
\usage{
\method{plot}{fit_result}(x, dyn = 1, x_pos = 1, y_pos = 1,
  z_pos = 1, coil = 1, xlim = NULL, data_only = FALSE,
  label = NULL, plot_sigs = NULL, n = NULL, sub_bl = FALSE,
  mar = NULL, restore_def_par = TRUE, ylim = NULL, y_scale = FALSE,
  ...)
}
\arguments{
\item{x}{fit_result object.}

\item{dyn}{the dynamic index to plot.}

\item{x_pos}{the x index to plot.}

\item{y_pos}{the y index to plot.}

\item{z_pos}{the z index to plot.}

\item{coil}{the coil element number to plot.}

\item{xlim}{the range of values to display on the x-axis, eg xlim = c(4,1).}

\item{data_only}{display only the processed data (logical).}

\item{label}{character string to add to the top left of the plot window.}

\item{plot_sigs}{a character vector of signal names to add to the plot.}

\item{n}{single index element to plot (overrides other indices when given).}

\item{sub_bl}{subtract the baseline from the data and fit (logical).}

\item{mar}{option to adjust the plot margins. See ?par.}

\item{restore_def_par}{restore default plotting par values after the plot has
been made.}

\item{ylim}{range of values to display on the y-axis, eg ylim = c(0,10).}

\item{y_scale}{option to display the y-axis values (logical).}

\item{...}{further arguments to plot method.}
}
\description{
Plot the fitting results of an object of class \code{fit_result}.
}
