% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{plot.mrs_data}
\alias{plot.mrs_data}
\title{Plotting method for objects of class mrs_data.}
\usage{
\method{plot}{mrs_data}(
  x,
  dyn = 1,
  x_pos = 1,
  y_pos = 1,
  z_pos = 1,
  coil = 1,
  fd = TRUE,
  x_units = NULL,
  xlim = NULL,
  y_scale = FALSE,
  x_ax = TRUE,
  mode = "re",
  lwd = NULL,
  bty = NULL,
  label = "",
  restore_def_par = TRUE,
  mar = NULL,
  xaxis_lab = NULL,
  xat = NULL,
  xlabs = TRUE,
  yat = NULL,
  ylabs = TRUE,
  ...
)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{dyn}{the dynamic index to plot.}

\item{x_pos}{the x index to plot.}

\item{y_pos}{the y index to plot.}

\item{z_pos}{the z index to plot.}

\item{coil}{the coil element number to plot.}

\item{fd}{display data in the frequency-domain (default), or time-domain
(logical).}

\item{x_units}{the units to use for the x-axis, can be one of: "ppm", "hz",
"points" or "seconds".}

\item{xlim}{the range of values to display on the x-axis, eg xlim = c(4,1).}

\item{y_scale}{option to display the y-axis values (logical).}

\item{x_ax}{option to display the x-axis values (logical).}

\item{mode}{representation of the complex numbers to be plotted, can be one
of: "re", "im", "mod" or "arg".}

\item{lwd}{plot linewidth.}

\item{bty}{option to draw a box around the plot. See ?par.}

\item{label}{character string to add to the top left of the plot window.}

\item{restore_def_par}{restore default plotting par values after the plot has
been made.}

\item{mar}{option to adjust the plot margins. See ?par.}

\item{xaxis_lab}{x-axis label.}

\item{xat}{x-axis tick label values.}

\item{xlabs}{x-axis tick labels.}

\item{yat}{y-axis tick label values.}

\item{ylabs}{y-axis tick labels.}

\item{...}{other arguments to pass to the plot method.}
}
\description{
Plotting method for objects of class mrs_data.
}
