% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{plot_slice_fit}
\alias{plot_slice_fit}
\title{Plot a 2D slice from an MRSI fit result object.}
\usage{
plot_slice_fit(fit_res, name, slice = 1, zlim = NULL, interp = 1)
}
\arguments{
\item{fit_res}{\code{fit_result} object.}

\item{name}{name of the quantity to plot, eg "TNAA".}

\item{slice}{slice to plot in the z direction.}

\item{zlim}{range of values to plot.}

\item{interp}{interpolation factor.}
}
\description{
Plot a 2D slice from an MRSI fit result object.
}
