% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{plot_slice_fit_inter}
\alias{plot_slice_fit_inter}
\title{Plot a 2D slice from an MRSI fit result object.}
\usage{
plot_slice_fit_inter(fit_res, map = NULL, slice = 1, zlim = NULL,
  interp = 1, xlim = NULL)
}
\arguments{
\item{fit_res}{\code{fit_result} object.}

\item{map}{array of values to be plotted, defaults to a "TNAA" map.}

\item{slice}{slice to plot in the z direction.}

\item{zlim}{range of values to plot.}

\item{interp}{interpolation factor.}

\item{xlim}{spectral plot limits for the x axis.}
}
\description{
Plot a 2D slice from an MRSI fit result object.
}
