% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plotting.R
\name{plot_slice_map_inter}
\alias{plot_slice_map_inter}
\title{Plot an interactive slice map from a data array where voxels can be selected
to display a corresponding spectrum.}
\usage{
plot_slice_map_inter(mrs_data, map = NULL, xlim = NULL, slice = 1,
  zlim = NULL, mask_map = NULL, denom = NULL, mask_cutoff = 20,
  interp = 1, mode = "re", y_scale = FALSE, ylim = NULL)
}
\arguments{
\item{mrs_data}{spectral data.}

\item{map}{array of values to be plotted, defaults to the integration of the
modulus of the full spectral width.}

\item{xlim}{spectral region to plot.}

\item{slice}{the slice index to plot.}

\item{zlim}{smallest and largest values to be plotted.}

\item{mask_map}{matching map with logical values to indicate if the
corresponding values should be plotted.}

\item{denom}{map to use as a denominator.}

\item{mask_cutoff}{minimum values to plot (as a percentage of the maximum).}

\item{interp}{map interpolation factor.}

\item{mode}{representation of the complex spectrum to be plotted, can be one
of: "re", "im", "mod" or "arg".}

\item{y_scale}{option to display the y-axis values (logical).}

\item{ylim}{intensity range to plot.}
}
\description{
Plot an interactive slice map from a data array where voxels can be selected
to display a corresponding spectrum.
}
