% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{plot_voi_overlay}
\alias{plot_voi_overlay}
\title{Plot a volume as an image overlay.}
\usage{
plot_voi_overlay(voi, mri)
}
\arguments{
\item{voi}{volume data as a nifti object.}

\item{mri}{image data as a nifti object.}
}
\description{
Plot a volume as an image overlay.
}
