% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{plot_voi_overlay}
\alias{plot_voi_overlay}
\title{Plot a volume as an image overlay.}
\usage{
plot_voi_overlay(voi, mri, flip_lr = FALSE, export_path = NULL)
}
\arguments{
\item{voi}{volume data as a nifti object or path to data file.}

\item{mri}{image data as a nifti object or path to data file.}

\item{flip_lr}{flip the image in the left-right direction.}

\item{export_path}{optional path to save the image in png format.}
}
\description{
Plot a volume as an image overlay.
}
