% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{plot_voi_overlay_seg}
\alias{plot_voi_overlay_seg}
\title{Plot a volume as an overlay on a segmented brain volume.}
\usage{
plot_voi_overlay_seg(voi, mri_seg, export_path = NULL, ...)
}
\arguments{
\item{voi}{volume data as a nifti object.}

\item{mri_seg}{segmented brain volume as a nifti object.}

\item{export_path}{optional path to save the image in png format.}

\item{...}{additional arguments to the ortho3 function.}
}
\description{
Plot a volume as an overlay on a segmented brain volume.
}
