% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmrs.R
\name{preproc_svs}
\alias{preproc_svs}
\title{Preprocess and perform quality assessment of a single SVS data set.}
\usage{
preproc_svs(
  path,
  label = NULL,
  output_dir = NULL,
  ref_inds = NULL,
  dyn_water_ref_path = NULL,
  Ntrans = NULL
)
}
\arguments{
\item{path}{path to the MRS data file or IMA directory.}

\item{label}{a label to describe the data set.}

\item{output_dir}{output directory.}

\item{ref_inds}{a vector of 1-based indices for any water reference dynamic
scans.}

\item{dyn_water_ref_path}{path to interleaved water reference data file or
IMA directory.}

\item{Ntrans}{override the number of transients set in the mrs_data file.}
}
\description{
Preprocess and perform quality assessment of a single SVS data set.
}
