% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_set.R
\name{read_basis}
\alias{read_basis}
\title{Read a basis file in LCModel .basis format.}
\usage{
read_basis(basis_file, ref = def_ref(), sort_basis = TRUE)
}
\arguments{
\item{basis_file}{path to basis file.}

\item{ref}{assumed ppm reference value.}

\item{sort_basis}{sort the basis set based on signal names.}
}
\value{
basis object.
}
\description{
Read a basis file in LCModel .basis format.
}
