% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{read_lcm_coord}
\alias{read_lcm_coord}
\title{Read an LCModel formatted coord file contaning fit information.}
\usage{
read_lcm_coord(coord_f)
}
\arguments{
\item{coord_f}{Path to the coord file.}
}
\value{
A list containg a table of fit point and results structure contaning
signal amplitudes, errors and fitting diagnostics.
}
\description{
Read an LCModel formatted coord file contaning fit information.
}
