% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulse_shapes.R
\name{read_pulse_pta}
\alias{read_pulse_pta}
\title{Read a .pta formatted text file compatible with Siemens PulseTool.}
\usage{
read_pulse_pta(fname)
}
\arguments{
\item{fname}{pta formatted pulse file name.}
}
\value{
pulse waveform and header.
}
\description{
Read a .pta formatted text file compatible with Siemens PulseTool.
}
