% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_read_twix.R
\name{read_siemens_txt_hdr}
\alias{read_siemens_txt_hdr}
\title{Read the text format header found in Siemens IMA and TWIX data files.}
\usage{
read_siemens_txt_hdr(input, version = "vd", verbose = FALSE, offset = 0)
}
\arguments{
\item{input}{file name to read or raw data.}

\item{version}{software version, can be "vb" or "vd".}

\item{verbose}{print information to the console.}

\item{offset}{offset to begin searching for the text header.}
}
\value{
a list of parameter values
}
\description{
Read the text format header found in Siemens IMA and TWIX data files.
}
