% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reg.R
\name{resample_img}
\alias{resample_img}
\title{Resample an image to match a target image space.}
\usage{
resample_img(source, target)
}
\arguments{
\item{source}{image data as a nifti object.}

\item{target}{image data as a nifti object.}
}
\value{
resampled image data as a nifti object.
}
\description{
Resample an image to match a target image space.
}
