% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{reson_table2mrs_data}
\alias{reson_table2mrs_data}
\title{Generate mrs_data from a table of single Lorentzian resonances.}
\usage{
reson_table2mrs_data(reson_table, acq_paras = def_acq_paras())
}
\arguments{
\item{reson_table}{as produced by the hsvd function.}

\item{acq_paras}{list of acquisition parameters. See
\code{\link{def_acq_paras}}}
}
\value{
mrs_data object.
}
\description{
Generate mrs_data from a table of single Lorentzian resonances.
}
