% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_scaling.R
\name{scale_amp_molal}
\alias{scale_amp_molal}
\title{Apply water reference scaling to a fitting results object to yield metabolite
quantities in millimolar (mM) units (mol / kg of tissue water).}
\usage{
scale_amp_molal(
  fit_result,
  ref_data,
  te,
  tr,
  water_t1,
  water_t2,
  metab_t1,
  metab_t2,
  water_conc,
  ...
)
}
\arguments{
\item{fit_result}{result object generated from fitting.}

\item{ref_data}{water reference MRS data object.}

\item{te}{the MRS TE in seconds.}

\item{tr}{the MRS TR in seconds.}

\item{water_t1}{assumed water T1 value.}

\item{water_t2}{assumed water T2 value.}

\item{metab_t1}{assumed metabolite T1 value.}

\item{metab_t2}{assumed metabolite T2 value.}

\item{water_conc}{assumed MR-visible water concentration in molal units
(moles / gram). For pure water this is 55510.0 (mol / g).}

\item{...}{additional arguments to get_td_amp function.}
}
\value{
A \code{fit_result} object with a rescaled results table.
}
\description{
Details of this method can be found in "Use of tissue water as a
concentration reference for proton spectroscopic imaging" by Gasparovic et al
MRM 2006 55(6):1219-26.
}
