% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_scaling.R
\name{scale_amp_molal_pvc}
\alias{scale_amp_molal_pvc}
\title{Apply partial volume correction to a fitting result object.}
\usage{
scale_amp_molal_pvc(fit_result, ref_data, p_vols, te, tr, ...)
}
\arguments{
\item{fit_result}{result object generated from fitting.}

\item{ref_data}{water reference MRS data object.}

\item{p_vols}{a numeric vector of partial volumes.}

\item{te}{the MRS TE in seconds.}

\item{tr}{the MRS TR in seconds.}

\item{...}{additional arguments to get_td_amp function.}
}
\value{
A \code{fit_result} object with a rescaled results table.
}
\description{
Apply partial volume correction to a fitting result object.
}
