% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sd}
\alias{sd}
\title{Calculate the standard deviation spectrum from an mrs_data object.}
\usage{
sd(x, na.rm)
}
\arguments{
\item{x}{object of class mrs_data.}

\item{na.rm}{remove NA values.}
}
\value{
sd mrs_data object.
}
\description{
Calculate the standard deviation spectrum from an mrs_data object.
}
