% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{shift}
\alias{shift}
\title{Apply a frequency shift to MRS data.}
\usage{
shift(mrs_data, shift, units = "ppm")
}
\arguments{
\item{mrs_data}{MRS data.}

\item{shift}{frequency shift (in ppm by default).}

\item{units}{of the shift ("ppm" or "hz").}
}
\value{
frequency shifted MRS data.
}
\description{
Apply a frequency shift to MRS data.
}
