% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_simulation.R
\name{sim_basis}
\alias{sim_basis}
\title{Simulate a basis set object.}
\usage{
sim_basis(mol_list, pul_seq = seq_pulse_acquire, ft = def_ft(),
  ref = def_ref(), fs = def_fs(), N = def_N(), xlim = NULL, ...)
}
\arguments{
\item{mol_list}{a list of \code{mol_parameter} objects.}

\item{pul_seq}{A pulse sequence function to use.}

\item{ft}{Transmitter frequency in Hz.}

\item{ref}{Reference value for ppm scale.}

\item{fs}{Sampling frequency in Hz.}

\item{N}{Number of data points in the spectral dimension.}

\item{xlim}{A ppm range limiting signals to be simulated.}

\item{...}{Extra parameters to pass to the pulse sequence function.}
}
\value{
A basis object.
}
\description{
Simulate a basis set object.
}
