% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sim_noise}
\alias{sim_noise}
\title{Simulate an mrs_data object containing simulated Gaussian noise.}
\usage{
sim_noise(
  sd = 0.1,
  fs = def_fs(),
  ft = def_ft(),
  N = def_N(),
  ref = def_ref(),
  nuc = def_nuc(),
  dyns = 1,
  fd = TRUE
)
}
\arguments{
\item{sd}{standard deviation of the noise.}

\item{fs}{sampling frequency in Hz.}

\item{ft}{transmitter frequency in Hz.}

\item{N}{number of data points in the spectral dimension.}

\item{ref}{reference value for ppm scale.}

\item{nuc}{resonant nucleus.}

\item{dyns}{number of dynamic scans to generate.}

\item{fd}{return data in the frequency-domain (TRUE) or time-domain (FALSE)}
}
\value{
mrs_data object.
}
\description{
Simulate an mrs_data object containing simulated Gaussian noise.
}
