% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spant.R
\docType{package}
\name{spant-package}
\alias{spant}
\alias{spant-package}
\title{spant: spectroscopy analysis tools.}
\description{
spant provides a set of tools for reading, visualising and processing
Magnetic Resonance Spectroscopy (MRS) data.
}
\details{
To learn more about spant, start with the vignettes:
\code{browseVignettes(package = "spant")}

For a full list of functions:
\code{help(package=spant,help_type="html")}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/martin3141/spant/issues}
}

}
\author{
\strong{Maintainer}: Martin Wilson \email{martin@pipegrep.co.uk}

}
