% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_display.R
\name{stackplot.fit_result}
\alias{stackplot.fit_result}
\title{Plot the fitting results of an object of class \code{fit_result} with
individual basis set components shown.}
\usage{
\method{stackplot}{fit_result}(x, xlim = c(4, 0.5), y_offset = 0.04,
  plt_title = FALSE, n = NA, ...)
}
\arguments{
\item{x}{fit_result object.}

\item{xlim}{the range of values to display on the x-axis, eg xlim = c(4,1).}

\item{y_offset}{seperate basis signals in the y-axis direction by this value.}

\item{plt_title}{title to add to the plot.}

\item{n}{index to the fit number to display.}

\item{...}{further arguments to plot method.}
}
\description{
Plot the fitting results of an object of class \code{fit_result} with
individual basis set components shown.
}
