% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_display.R
\name{stackplot.mrs_data}
\alias{stackplot.mrs_data}
\title{TODO make consistant with plot and document.}
\usage{
\method{stackplot}{mrs_data}(x, mode = "real", xlim = NULL, x_offset = 0,
  y_offset = 5, dim = "dyn", x_pos = NA, y_pos = NA, z_pos = NA,
  dyn = 1, coil = 1, ...)
}
\description{
TODO make consistant with plot and document.
}
