% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{sub_mean_dyns}
\alias{sub_mean_dyns}
\title{Subtract the mean dynamic spectrum from a dynamic series.}
\usage{
sub_mean_dyns(mrs_data, scale = 1)
}
\arguments{
\item{mrs_data}{dynamic MRS data.}

\item{scale}{scale factor for the mean spectrum.}
}
\value{
subtracted data.
}
\description{
Subtract the mean dynamic spectrum from a dynamic series.
}
