% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svs_batch_fit.R
\name{svs_1h_brain_analysis}
\alias{svs_1h_brain_analysis}
\title{Standard SVS 1H brain analysis pipeline.}
\usage{
svs_1h_brain_analysis(
  metab,
  basis = NULL,
  w_ref = NULL,
  mri_seg = NULL,
  mri = NULL,
  output_dir = NULL,
  extra = NULL,
  decimate = NULL,
  rats_corr = TRUE,
  ecc = FALSE,
  comb_dyns = TRUE,
  hsvd_filt = FALSE,
  scale_amps = TRUE,
  te = NULL,
  tr = NULL,
  preproc_only = FALSE
)
}
\description{
Standard SVS 1H brain analysis pipeline.
}
