% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_io.R
\name{write_mrs}
\alias{write_mrs}
\title{Write MRS data object to file.}
\usage{
write_mrs(fname, mrs_data, format = NULL)
}
\arguments{
\item{fname}{the filename of the output.}

\item{mrs_data}{object to be written to file.}

\item{format}{string describing the data format. Must be one of the
following : "nifti", "dpt", "lcm_raw", "rds". If not specified, the format
will be guessed from the filename extension.}
}
\description{
Write MRS data object to file.
}
