% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zero_spec_threshold}
\alias{zero_spec_threshold}
\title{Set all spectral data points below a given percentage of the maximum to zero.}
\usage{
zero_spec_threshold(mrs_data, percent_max)
}
\arguments{
\item{mrs_data}{data to be processed.}
}
\value{
mrs_data object with zeroed spectral data points.
}
\description{
Set all spectral data points below a given percentage of the maximum to zero.
}
