% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zf}
\alias{zf}
\title{Zero-fill MRS data in the time domain.}
\usage{
zf(mrs_data, factor = 2)
}
\arguments{
\item{mrs_data}{MRS data.}

\item{factor}{Zero-filling factor, factor of 2 returns a dataset with
twice the original data points.}
}
\value{
Zero-filled data.
}
\description{
Zero-fill MRS data in the time domain.
}
