% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zf}
\alias{zf}
\alias{zf.mrs_data}
\alias{zf.basis_set}
\title{Zero-fill MRS data in the time domain.}
\usage{
zf(x, factor = 2)

\method{zf}{mrs_data}(x, factor = 2)

\method{zf}{basis_set}(x, factor = 2)
}
\arguments{
\item{x}{input mrs_data or basis_set object.}

\item{factor}{zero-filling factor, factor of 2 returns a dataset with
twice the original data points.}
}
\value{
zero-filled data.
}
\description{
Zero-fill MRS data in the time domain.
}
