% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_data_proc.R
\name{zf_xy}
\alias{zf_xy}
\title{Zero-fill MRSI data in the k-space x-y direction.}
\usage{
zf_xy(mrs_data, factor = 2)
}
\arguments{
\item{mrs_data}{MRSI data.}

\item{factor}{Zero-filling factor, factor of 2 returns a dataset with
twice the orignal points in the x-y directions.}
}
\value{
Zero-filled data.
}
\description{
Zero-fill MRSI data in the k-space x-y direction.
}
